% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny-modules-new.R
\name{accountPicker}
\alias{accountPicker}
\title{accountPicker [Shiny Module] - pick GA4 accounts/webProperties in Shiny}
\usage{
accountPicker(id, ga_table, id_only = TRUE)
}
\arguments{
\item{ga_table}{A table GA4 accounts/web properties from \code{ga_account_summary("ga4")}}

\item{id_only}{Whether to return just the id, not the row}
}
\value{
If id_only=FALSE, the row of ga_table for the selected GA4 web property e.g. use \code{ga_table$propertyId} to send to \link{ga_data} calls.  If id_only=TRUE, just the propertyId
}
\description{
Shiny Module for use with \link{accountPickerUI}
}
\details{
Call via \code{accountPicker("your_id", ga_table)}
}
\seealso{
Other Shiny modules: 
\code{\link{accountPickerUI}()},
\code{\link{authDropdownUI}()},
\code{\link{authDropdown}()},
\code{\link{multi_selectUI}()},
\code{\link{multi_select}()}
}
\concept{Shiny modules}
