% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate.R
\name{aggregateGAData}
\alias{aggregateGAData}
\title{Aggregate a Google Analytics dataframe over inputted columns}
\usage{
aggregateGAData(ga_data, agg_names = NULL,
  mean_regex = "^avg|^percent|Rate$|^CPC$|^CTR$|^CPM$|^RPC$|^ROI$|^ROAS$|Per")
}
\arguments{
\item{ga_data}{A dataframe of data to aggregate}

\item{agg_names}{The columns to aggregate over}

\item{mean_regex}{The regex for column names to do mean() rather than sum()}
}
\description{
A helper function to aggregate over dimensions
}
\details{
Will auto select metrics if they are numeric class columns.
  Will auto perform mean aggregation it metric names match \code{mean_regex} argument
  If agg_names is NULL will aggregate over all
}
