% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom_dims_mets.R
\name{allowed_metric_dim}
\alias{allowed_metric_dim}
\title{Create named list of allowed GA metrics/dimensions}
\usage{
allowed_metric_dim(type = c("METRIC", "DIMENSION"), subType = c("all",
  "segment", "cohort"), callAPI = FALSE)
}
\arguments{
\item{type}{Type of parameter to create}

\item{subType}{to restrict to only those in this type}

\item{callAPI}{This will update the meta table (Requires online authorization)

This is useful to expand goalXCompletions to all the possibilities,
  as well as restricting to those that variables that work with your API call.

Use internal meta table, but you have option to update to the latest version.}
}
\value{
A named list of parameters for use in API calls
}
\description{
Create named list of allowed GA metrics/dimensions
}
