% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{assign_list_class}
\alias{assign_list_class}
\title{when we need a list of objs of class x to also be class x}
\usage{
assign_list_class(x, the_class)
}
\description{
when we need a list of objs of class x to also be class x
}
