% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny-modules.R
\name{authDropdown}
\alias{authDropdown}
\title{authDropdown [Shiny Module]}
\usage{
authDropdown(input, output, session, ga.table, viewIdOnly = TRUE, rmNA = TRUE)
}
\arguments{
\item{input}{shiny input}

\item{output}{shiny output}

\item{session}{shiny session}

\item{ga.table}{A table of GA tables}

\item{viewIdOnly}{Default only returns the viewId, set to FALSE to return the row of ga.table satisfying the selections}

\item{rmNA}{Will remove any rows that have NA listed for the columns.  Set to FALSE to return all rows.}
}
\value{
GA View Id selected
}
\description{
Shiny Module for use with \link{authDropdownUI}
}
\details{
Call via \code{shiny::callModule(authDropdown, "your_id")}
}
\seealso{
Other Shiny modules: 
\code{\link{accountPickerUI}()},
\code{\link{authDropdownUI}()},
\code{\link{multi_selectUI}()},
\code{\link{multi_select}()}
}
\concept{Shiny modules}
