% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny-modules.R
\name{authDropdownUI}
\alias{authDropdownUI}
\title{authDropdown UI}
\usage{
authDropdownUI(id, width = NULL, inColumns = FALSE)
}
\arguments{
\item{id}{Shiny id}

\item{width}{The width of the input, e.g. '400px', or '100%'}

\item{inColumns}{whether to wrap selectInputs in width=4 columns.}
}
\value{
Shiny UI
}
\description{
Shiny Module for use with \link{authDropdown}
}

