% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_web_objects.R, R/ga4_data_objects.R
\name{BatchRunReportsRequest}
\alias{BatchRunReportsRequest}
\title{BatchRunReportsRequest Object}
\usage{
BatchRunReportsRequest(entity, requests)

BatchRunReportsRequest(entity, requests)
}
\arguments{
\item{entity}{A property whose events are tracked}

\item{requests}{Individual requests}
}
\value{
BatchRunReportsRequest object

BatchRunReportsRequest object
}
\description{
BatchRunReportsRequest Object

BatchRunReportsRequest Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The batch request containing multiple report requests.

Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The batch request containing multiple report requests.
}
\concept{BatchRunReportsRequest functions}
