% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ga4_data_objects_filters.R
\name{BetweenFilter}
\alias{BetweenFilter}
\title{BetweenFilter Object}
\usage{
BetweenFilter(toValue, fromValue)
}
\arguments{
\item{toValue}{Ends with this number}

\item{fromValue}{Begins with this number}
}
\value{
BetweenFilter object
}
\description{
BetweenFilter Object
}
\details{
To express that the result needs to be between two numbers (inclusive).
}
\concept{BetweenFilter functions}
