% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{checkPrefix}
\alias{checkPrefix}
\title{check it starts with ga: and if not puts it on}
\usage{
checkPrefix(x, prefix = c("ga", "mcf"))
}
\description{
check it starts with ga: and if not puts it on
}
\keyword{internal}
