% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anti_sample.R
\name{chunkify}
\alias{chunkify}
\title{Break down a request into unsampled chunks}
\usage{
chunkify(sessions_vec, limit = 250000)
}
\arguments{
\item{sessions_vec}{A vector, ordered by date, with the number of sessions}

\item{limit}{This is the upper bound for the number of sessions in one chunk}
}
\value{
A vector with the batch number of each date. This allows for usage in
  a mutate.
}
\description{
Break down a request into unsampled chunks
}
\keyword{internal}
