% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cohorts.R
\name{cohort}
\alias{cohort}
\title{Cohort object}
\usage{
cohort(name, type = c("FIRST_VISIT_DATE"), dateRange)
}
\arguments{
\item{name}{A unique name for the cohort}

\item{type}{Type of cohort.  Only FIRST_VISIT_DATE is support at the moment.}

\item{dateRange}{A \link{date_ga4} object}
}
\description{
Cohort object
}
\seealso{
Other v4 cohort functions: \code{\link{cohortGroup}},
  \code{\link{cohort_dimension_check}},
  \code{\link{cohort_metric_check}},
  \code{\link{make_cohort_group}}
}
\keyword{internal}
