% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cohorts.R
\name{cohort_dimension_check}
\alias{cohort_dimension_check}
\title{Check the request has cohort dimensions}
\usage{
cohort_dimension_check(dimensions)
}
\arguments{
\item{metrics}{A vector of dimensions}
}
\description{
Check the request has cohort dimensions
}
\seealso{
Other v4 cohort functions: \code{\link{cohortGroup}},
  \code{\link{cohort_metric_check}}, \code{\link{cohort}},
  \code{\link{make_cohort_group}}
}
\keyword{internal}
