% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cohorts.R
\name{cohort_metric_check}
\alias{cohort_metric_check}
\title{Check the request has cohort metrics}
\usage{
cohort_metric_check(metrics)
}
\arguments{
\item{metrics}{A vector of metrics}
}
\description{
Check the request has cohort metrics
}
\seealso{
Other v4 cohort functions: \code{\link{cohortGroup}},
  \code{\link{cohort_dimension_check}},
  \code{\link{cohort}}, \code{\link{make_cohort_group}}
}
\keyword{internal}
