% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cohorts.R
\name{cohortGroup}
\alias{cohortGroup}
\title{Cohort Group object}
\usage{
cohortGroup(cohorts, lifetimeValue = FALSE)
}
\arguments{
\item{cohorts}{A list of cohort objects}

\item{lifetimeValue}{Boolean. Has to be an app.}
}
\description{
Cohort Group object
}
\seealso{
Other v4 cohort functions: \code{\link{cohort_dimension_check}},
  \code{\link{cohort_metric_check}}, \code{\link{cohort}},
  \code{\link{make_cohort_group}}
}
\keyword{internal}
