% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom_dims_mets.R
\name{customDimension}
\alias{customDimension}
\title{A customDimension object}
\usage{
customDimension(name = NULL, index = NULL, scope = NULL, active = NULL,
  id = NULL, accountId = NULL, webPropertyId = NULL)
}
\arguments{
\item{name}{Name of custom dimension}

\item{index}{Index of custom dimension - integer between 1 and 20 (200 for GA360)}

\item{scope}{Scope of custom dimension - one of \code{"HIT","SESSION","USER","PRODUCT"}}

\item{active}{\code{TRUE} or \code{FALSE} if custom dimension is active or not}

\item{id}{The id of the custom dimension}

\item{accountId}{AccountId of the custom dimension}

\item{webPropertyId}{WebPropertyId of the custom dimension}
}
\description{
Pass this into \link{ga_custom_vars_create} or \link{ga_custom_var_update}
}
\details{
When creating a new custom dimension, supply all the arguments except id.  
 
 If you are updating an existing custom dimension, only supply the arguments you wish to change
}
\seealso{
Other custom variable functions: \code{\link{ga_custom_vars_create}},
  \code{\link{ga_custom_vars_list}},
  \code{\link{ga_custom_vars_patch}},
  \code{\link{ga_custom_vars}}
}
