% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ga_v4_objects.R
\name{dim_filter}
\alias{dim_filter}
\title{Make a dimension filter object}
\usage{
dim_filter(dimension, operator = c("REGEXP", "BEGINS_WITH", "ENDS_WITH",
  "PARTIAL", "EXACT", "NUMERIC_EQUAL", "NUMERIC_GREATER_THAN",
  "NUMERIC_LESS_THAN", "IN_LIST"), expressions, caseSensitive = FALSE,
  not = FALSE)
}
\arguments{
\item{dimension}{dimension name to filter on.}

\item{operator}{How to match the dimension.}

\item{expressions}{What to match.  A character vector if operator is "IN_LIST"}

\item{caseSensitive}{Boolean.}

\item{not}{Logical NOT operator. Boolean.}
}
\value{
An object of class \code{dim_fil_ga4} for use in \code{\link{filter_clause_ga4}}
}
\description{
Make a dimension filter object
}
\seealso{
Other filter functions: \code{\link{filter_clause_ga4}},
  \code{\link{met_filter}}
}

