% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_web_objects.R
\name{Dimension}
\alias{Dimension}
\title{Dimension Object}
\usage{
Dimension(name, dimensionExpression = NULL)
}
\arguments{
\item{name}{The name of the dimension}

\item{dimensionExpression}{One dimension can be the result of an expression of multiple dimensions}
}
\value{
Dimension object
}
\description{
Dimension Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Dimensions are attributes of your data. For example, the dimension City indicates the city, for example, 'Paris' or 'New York', from which an event originates. Requests are allowed up to 8 dimensions.
}
\concept{Dimension functions}
