% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ga_v4_objects.R
\name{dimension_ga4}
\alias{dimension_ga4}
\title{Make a dimension object}
\usage{
dimension_ga4(vector, histogramBuckets = NULL)
}
\arguments{
\item{vector}{A character vector of dimension names}

\item{histogramBuckets}{Optional named list of histogram buckets vectors}
}
\value{
a list suitable for parsing in req
}
\description{
Make a dimension object
}
\section{histograms}{


 If supplying histogramBuckets, the name will be used to create the buckets

 e.g. \code{dimensions = c("source", "medium")
            histogramBuckets = list(source = c(1,5,10,20,200),
                                   medium = c(0,7,14,21,999))
                                   }

 Warning: only makes sense to apply to dimensions that are numeric e.g. ga:hour
}

\keyword{internal}
