% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ga_data_order.R
\name{DimensionOrderBy}
\alias{DimensionOrderBy}
\title{DimensionOrderBy Object}
\usage{
DimensionOrderBy(
  dimensionName,
  orderType = c("ALPHANUMERIC", "CASE_INSENSITIVE_ALPHANUMERIC", "NUMERIC")
)
}
\arguments{
\item{dimensionName}{A dimension name in the request to order by}

\item{orderType}{Controls the rule for dimension value ordering}
}
\value{
DimensionOrderBy object
}
\description{
DimensionOrderBy Object
}
\details{
Sorts by dimension values.

\code{orderType} has values:

Alphanumeric sort by Unicode code point. For example, "2" < "A" < "X" < "b" < "z".

Case insensitive alphanumeric sort by lower case Unicode code point. For example, "2" < "A" < "b" < "X" < "z".

Numeric: Dimension values are converted to numbers before sorting. For example in NUMERIC sort, "25" < "100", and in ALPHANUMERIC sort, "100" < "25". Non-numeric dimension values all have equal ordering value below all numeric values.
}
\concept{DimensionOrderBy functions}
