% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ga_v4_objects.R
\name{dynamicSegment}
\alias{dynamicSegment}
\title{Dynamic Segment v4 object}
\usage{
dynamicSegment(name, userSegment, sessionSegment)
}
\description{
Dynamic Segment v4 object
}
\seealso{
Other v4 segment functions: \code{\link{orFiltersForSegment}},
  \code{\link{segmentDefinition}},
  \code{\link{segmentDimensionFilter}},
  \code{\link{segmentFilterClause}},
  \code{\link{segmentFilter}},
  \code{\link{segmentMetricFilter}},
  \code{\link{segmentObj_ga4}},
  \code{\link{segmentSequenceStep}},
  \code{\link{segment_filter}}, \code{\link{segment_ga4}},
  \code{\link{sequenceSegment}},
  \code{\link{simpleSegment}}
}

