% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_web_objects.R
\name{Entity}
\alias{Entity}
\title{Entity Object}
\usage{
Entity(propertyId = NULL)
}
\arguments{
\item{propertyId}{A Google Analytics App + Web property id}
}
\value{
Entity object
}
\description{
Entity Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The unique identifier of the property whose events are tracked.
}
\concept{Entity functions}
