% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{error.message}
\alias{error.message}
\title{Get the error message}
\usage{
error.message(test_me)
}
\arguments{
\item{test_me}{an object that has failed is.error}
}
\value{
The error message
}
\description{
Get the error message
}
\keyword{internal}
