% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{expect_list_of_class}
\alias{expect_list_of_class}
\title{Test S3 Class in a list}
\usage{
expect_list_of_class(listthing, types, null_ok = FALSE)
}
\arguments{
\item{listthing}{A list of things}

\item{types}{A vector of s3 class we want}

\item{null_ok}{Is it ok to have a NULL listhing?}
}
\description{
Test S3 Class in a list
}
\keyword{internal}

