% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{expect_list_of_type}
\alias{expect_list_of_type}
\title{Test Type in a list}
\usage{
expect_list_of_type(listthing, types, null_ok = FALSE)
}
\arguments{
\item{listthing}{A list of things}

\item{types}{A vector of types we want}

\item{null_ok}{Is it ok to have a NULL listhing?}
}
\description{
Test Type in a list
}
\keyword{internal}

