% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{expect_null_or_s3_class}
\alias{expect_null_or_s3_class}
\title{Expect NULL or class (s3)}
\usage{
expect_null_or_s3_class(thing, s3class)
}
\description{
wraps assertthat::assert_that(inherits(thing, s3class))() to run if not NULL
}
\keyword{internal}
