% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{expect_null_or_type}
\alias{expect_null_or_type}
\title{Expect NULL or type}
\usage{
expect_null_or_type(thing, type)
}
\description{
wraps testthat::expect_type() to run if not NULL
}
\keyword{internal}
