% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ga_v4_get.R
\name{fetch_google_analytics_4_slow}
\alias{fetch_google_analytics_4_slow}
\title{Fetch GAv4 requests one at a time}
\usage{
fetch_google_analytics_4_slow(request_list, max_rows, allRows = FALSE)
}
\arguments{
\item{request_list}{A list of requests created by \link{make_ga_4_req}}

\item{max_rows}{Number of rows requested (if not fetched)}

\item{allRows}{Whether to fetch all available rows}
}
\value{
A dataframe of all the requests
}
\description{
Due to large complicated queries causing the v4 API to timeout, 
  this option is added to fetch via the more traditional one report per request
}
\seealso{
Other GAv4 fetch functions: \code{\link{fetch_google_analytics_4}},
  \code{\link{google_analytics_4}},
  \code{\link{make_ga_4_req}}
}
