% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_web_objects_filters.R
\name{Filter}
\alias{Filter}
\title{Filter Object}
\usage{
Filter(
  fieldName,
  numericFilter = NULL,
  inListFilter = NULL,
  nullFilter = NULL,
  stringFilter = NULL,
  betweenFilter = NULL
)
}
\arguments{
\item{fieldName}{The dimension name or metric name}

\item{numericFilter}{A filter for numeric or date values}

\item{inListFilter}{A filter for in list values}

\item{nullFilter}{A filter for null values}

\item{stringFilter}{Strings related filter}

\item{betweenFilter}{A filter for two values}
}
\value{
Filter object
}
\description{
Filter Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
An expression to filter dimension or metric values.
}
\concept{Filter functions}
