% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filters.R
\name{filter_clause_ga4}
\alias{filter_clause_ga4}
\title{Make a dimension or metric filter clause object}
\usage{
filter_clause_ga4(filters, operator = c("OR", "AND"))
}
\arguments{
\item{filters}{a list of \link{dim_filter} or \link{met_filter}. Only one type allowed.}

\item{operator}{combination of filter.}
}
\value{
An object of class \code{dim_fil_ga4} or \code{met_fil_ga4} for use in \code{\link{make_ga_4_req}}
}
\description{
Make a dimension or metric filter clause object
}
\details{
If you have dimension and metric filters,
  make the clauses in two seperate calls, then pass the objects to \link{make_ga_4_req}
}
\seealso{
Other filter functions: \code{\link{dim_filter}},
  \code{\link{met_filter}}
}

