% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ga4_data_objects_filters.R
\name{FilterExpression}
\alias{FilterExpression}
\title{FilterExpression Object}
\usage{
FilterExpression(
  andGroup = NULL,
  filter = NULL,
  orGroup = NULL,
  notExpression = NULL
)
}
\arguments{
\item{andGroup}{The FilterExpressions in and_group have an AND relationship}

\item{filter}{A primitive filter}

\item{orGroup}{The FilterExpressions in or_group have an OR relationship}

\item{notExpression}{The FilterExpression is NOT of not_expression}
}
\value{
FilterExpression object
}
\description{
FilterExpression Object
}
\details{
The fields in the same FilterExpression need to be either all dimensions or all metrics.
}
\concept{FilterExpression functions}
