% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analytics_admin.R
\name{ga_admin_account_summary}
\alias{ga_admin_account_summary}
\title{Returns summaries of all accounts accessible by the caller.}
\usage{
ga_admin_account_summary()
}
\description{
Returns summaries of all accounts accessible by the caller.
}
\details{

}
\seealso{
\href{http://code.google.com/apis/analytics/docs/mgmt/home.html}{Google Documentation}
}
