% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adwords.R
\name{ga_adwords}
\alias{ga_adwords}
\title{List AdWords}
\usage{
ga_adwords(accountId, webPropertyId)
}
\arguments{
\item{accountId}{Account Id}

\item{webPropertyId}{Web Property Id}
}
\value{
AdWords Links
}
\description{
List AdWords
}

