% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adwords.R
\name{ga_adwords_list}
\alias{ga_adwords_list}
\title{List AdWords}
\usage{
ga_adwords_list(accountId, webPropertyId)
}
\arguments{
\item{accountId}{Account Id}

\item{webPropertyId}{Web Property Id}
}
\value{
AdWords Links
}
\description{
List AdWords
}
\seealso{
Other Google Ad management functions: 
\code{\link{ga_adwords_add_linkid}()},
\code{\link{ga_adwords_delete_linkid}()},
\code{\link{ga_adwords}()}
}
\concept{Google Ad management functions}
