% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auth.R
\name{ga_auth}
\alias{ga_auth}
\title{ga_auth}
\usage{
ga_auth(new_user = FALSE, token = NULL)
}
\arguments{
\item{new_user}{Go through the auth flow again}

\item{token}{If you have an existing token already}
}
\description{
ga_auth
}

