% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ga_data_filter.R
\name{ga_aw_filter}
\alias{ga_aw_filter}
\title{Create a filter for use within App+Web filter expressions}
\usage{
ga_aw_filter(
  field,
  value,
  operation = c("EXACT", "BEGINS_WITH", "ENDS_WITH", "CONTAINS", "FULL_REGEXP",
    "PARTIAL_REGEXP", "EQUAL", "LESS_THAN", "LESS_THAN_OR_EQUAL", "GREATER_THAN",
    "GREATER_THAN_OR_EQUAL"),
  caseSensitive = TRUE
)
}
\arguments{
\item{field}{The dimension or metric field name}

\item{value}{The value - filter type is based on what class of object}

\item{operation}{How the filter operates on field with value}

\item{caseSensitive}{For character based value filters}
}
\description{
Used within \link{ga_aw_filter_expr}
}
\details{
Ensure your value is of the right R class for the type of filter you desire.

\itemize{
  \item{character: string filter}
  \item{character vector: in list filter}
  \item{numeric: Numeric filter}
  \item{numeric 2-length vector: between filter}
  \item{logical: TRUE will filter for NULLs}
 }

For numerics also make sure to specify integer or float for metrics e.g.
1L vs 1.00

Ensure operation is valid for the type of field you are passing:

string filters: "EXACT","BEGINS_WITH","ENDS_WITH","CONTAINS","FULL_REGEXP",
  "PARTIAL_REGEXP"
  
numeric filters: "EQUAL", "LESS_THAN", "LESS_THAN_OR_EQUAL", "GREATER_THAN",
  "GREATER_THAN_OR_EQUAL")
}
