% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_web_functions.R
\name{ga_aw_filter_expr}
\alias{ga_aw_filter_expr}
\title{Create a filter expression for use with App+Web Reports}
\usage{
ga_aw_filter_expr(..., type = c("default", "not", "and", "or"))
}
\arguments{
\item{...}{Vectors of \link{ga_aw_filter}, or vectors of FilterExpression created from previous calls to this function.}

\item{type}{The type of filter}
}
\value{
A FilterExpression
}
\description{
For use with \link{google_analytics_aw}
}
\details{
The fields in a FilterExpression need to be either all dimensions or all metrics. Use them in the appropriate dimensionFilter or metricFilter arguments within \link{google_analytics_aw}
}
\examples{

simple <- ga_aw_filter_expr(ga_aw_filter("city", "Copenhagen", "EXACT"))

multiple <- ga_aw_filter_expr(
     ga_aw_filter("city", "C", "BEGINS_WITH"),
     ga_aw_filter("continent", "Europe"),
   type = "and")
   
negative <- ga_aw_filter_expr(
     ga_aw_filter("city", "Copenhagen", "EXACT"),
     ga_aw_filter("date", c("20200416","20200415")),
   type = "not")

# use previously created FilterExpressions
complex <- ga_aw_filter_expr(

)
   

}
