% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{ga_cache_call}
\alias{ga_cache_call}
\title{Setup caching of API calls}
\usage{
ga_cache_call(cache_location)
}
\arguments{
\item{cache_location}{If \code{RAM} will save to memory, or specify a file folder location}
}
\description{
Lets you cache API calls to disk
}
\details{
By default this is turned on upon package load to \code{RAM}.  
  Should you want to cache calls to a folder then run this function to specify where.
}
