% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/client_id.R
\name{ga_clientid_hash}
\alias{ga_clientid_hash}
\title{Get hashed version of client id (also known as hashClientId, hashedClientId, or BigQuery's fullVisitorId)}
\usage{
ga_clientid_hash(webPropertyId, clientId)
}
\arguments{
\item{webPropertyId}{Web Property Id}

\item{clientId}{Client Id}
}
\value{
hashedClientId object list
}
\description{
Get hashed version of client id (also known as hashClientId, hashedClientId, or BigQuery's fullVisitorId)
}
\seealso{
Other clientid functions: 
\code{\link{ga_clientid_activity_unnest}()},
\code{\link{ga_clientid_activity}()},
\code{\link{ga_clientid_deletion}()}
}
\concept{clientid functions}
