% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom_data_source.R
\name{ga_custom_upload}
\alias{ga_custom_upload}
\title{Get Custom Data Source Upload Meta data}
\usage{
ga_custom_upload(accountId, webPropertyId, customDataSourceId, uploadId)
}
\arguments{
\item{accountId}{Account Id}

\item{webPropertyId}{Web Property Id}

\item{customDataSourceId}{Custom data source Id}

\item{uploadId}{upload Id}
}
\value{
Custom Data Source Upload Meta data
}
\description{
Get Custom Data Source Upload Meta data
}

