% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom_data_source.R
\name{ga_custom_upload_delete}
\alias{ga_custom_upload_delete}
\title{Deletes custom upload files for a given ids vector}
\usage{
ga_custom_upload_delete(accountId, webPropertyId, customDataSourceId,
  customDataImportUids)
}
\arguments{
\item{accountId}{Account Id}

\item{webPropertyId}{Web Property Id}

\item{customDataSourceId}{Custom data source Id}

\item{customDataImportUids}{vector of file upload ids.}
}
\description{
Deletes custom upload files for a given ids vector
}
\seealso{
https://developers.google.com/analytics/devguides/config/mgmt/v3/mgmtReference/management/uploads/deleteUploadData

Other custom datasource functions: \code{\link{ga_custom_datasource}},
  \code{\link{ga_custom_upload_file}},
  \code{\link{ga_custom_upload_list}},
  \code{\link{ga_custom_upload}}
}
\concept{custom datasource functions}
