% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom_data_source.R
\name{ga_custom_upload_file}
\alias{ga_custom_upload_file}
\title{Upload data to Google Analytics}
\usage{
ga_custom_upload_file(accountId, webPropertyId, customDataSourceId, upload)
}
\arguments{
\item{accountId}{Account Id}

\item{webPropertyId}{Web Property Id}

\item{customDataSourceId}{Custom data source Id}

\item{upload}{An R data.frame or a file path location (character).}
}
\value{
A custom upload Id object
}
\description{
Upload external data up to 1GB to Google Analytics via the management API.
}
\details{
You need to create a custom data source in the web UI first.  

If you are uploading an R data frame, the function will prefix the column names with \code{"ga:"} for you if necessary.
  
After upload check the status by querying data sources using \link{ga_custom_upload} 
  and examining the \code{status} field.

Currently only supports simple uploads (not resumable).
}
\seealso{
A guide for preparing the data is available \href{from Google here}{https://developers.google.com/analytics/solutions/data-import-cost}.

See also \href{Data Import Developer Guide}{https://developers.google.com/analytics/devguides/config/mgmt/v3/data-import#format}
}

