% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom_data_source.R
\name{ga_custom_uploads_list}
\alias{ga_custom_uploads_list}
\title{List Custom Data Uploads}
\usage{
ga_custom_uploads_list(accountId, webPropertyId, customDataSourceId)
}
\arguments{
\item{accountId}{Account Id}

\item{webPropertyId}{Web Property Id}

\item{customDataSourceId}{Custom data source Id}
}
\value{
Custom Data Source Uploads List
}
\description{
List Custom Data Uploads
}

