% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom_dims_mets.R
\name{ga_custom_vars_list}
\alias{ga_custom_vars_list}
\title{List Custom Dimensions or Metrics}
\usage{
ga_custom_vars_list(accountId, webPropertyId, type = c("customMetrics",
  "customDimensions"))
}
\arguments{
\item{accountId}{Account Id}

\item{webPropertyId}{Web Property Id}

\item{type}{A customMetric or customDimension}
}
\value{
AdWords Links
}
\description{
List Custom Dimensions or Metrics
}

