% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ga_data_funnel.R
\name{ga_data_funnel}
\alias{ga_data_funnel}
\title{Run a GA4 Data Funnel report}
\usage{
ga_data_funnel(
  funnel_steps,
  propertyId,
  open_funnel = FALSE,
  breakdown_dim = NULL,
  breakdown_limit = 5L,
  next_action = NULL,
  visual = "STANDARD_FUNNEL",
  dim_filters = NULL,
  limit = 10000L
)
}
\arguments{
\item{funnel_steps}{A list of FunnelStep objects}

\item{propertyId}{A GA4 property Id}

\item{open_funnel}{In an open funnel, users can enter the funnel in any step, and in a closed funnel, users must enter the funnel in the first step. Optional. If unspecified, a closed funnel is used.}

\item{dim_filters}{Filter on the dimensions of the request - a filter object created by \link{ga_data_filter}}

\item{limit}{The number of rows to return - use -1 to return all rows}
}
\description{
Returns a customized funnel report of your Google Analytics event data. The data returned from the API is as a table with columns for the requested dimensions and metrics.
}
\seealso{
Other GA4 functions: 
\code{\link{ga_data_filter}()},
\code{\link{ga_data_funnel_step}()},
\code{\link{ga_data_order}()},
\code{\link{ga_data}()}
}
\concept{GA4 functions}
