% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ga_data_funnel.R
\name{ga_data_funnel_step}
\alias{ga_data_funnel_step}
\title{Create a FunnelStep object for use within ga_data_funnel}
\usage{
ga_data_funnel_step(fun_filter, name, within_duration, directly_followed)
}
\arguments{
\item{fun_filter}{A funnel filter expression - the condition that your users must meet to be included in this step of the funnel journey.}

\item{name}{The distinctive name for this step. If unspecified, steps will be named by a 1 based indexed name (i.e. "0. ", "1. ", etc.).}

\item{within_duration}{If specified, this step must complete within this duration of the completion of the prior step.}

\item{directly_followed}{If true, this step must directly follow the previous step}
}
\description{
Create a FunnelStep object for use within ga_data_funnel
}
\seealso{
Other GA4 functions: 
\code{\link{ga_data_filter}()},
\code{\link{ga_data_funnel}()},
\code{\link{ga_data_order}()},
\code{\link{ga_data}()}
}
\concept{GA4 functions}
