% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ga_data_pivot.R
\name{ga_data_pivot}
\alias{ga_data_pivot}
\title{Create a GA4 Pivot object for the Data API}
\usage{
ga_data_pivot(
  fields,
  orderBys = NULL,
  offset = 0,
  limit = 10,
  metricAggregations = NULL
)
}
\description{
Create a GA4 Pivot object for the Data API
}
