% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experiments.R
\name{ga_experiment}
\alias{ga_experiment}
\title{Experiments Meta data}
\usage{
ga_experiment(accountId, webPropertyId, profileId, experimentId)
}
\arguments{
\item{accountId}{Account Id}

\item{webPropertyId}{Web Property Id}

\item{profileId}{Profile Id}

\item{experimentId}{Experiment Id}
}
\value{
Experiment Meta Data
}
\description{
Experiments Meta data
}
\seealso{
Other managementAPI functions: 
\code{\link{ga_experiment_list}()},
\code{\link{ga_filter_add}()},
\code{\link{ga_filter_apply_to_view}()},
\code{\link{ga_filter_update_filter_link}()},
\code{\link{ga_filter_update}()},
\code{\link{ga_segment_list}()}
}
\concept{managementAPI functions}
