% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experiments.R
\name{ga_experiment_list}
\alias{ga_experiment_list}
\title{List Experiments}
\usage{
ga_experiment_list(accountId, webPropertyId, profileId)
}
\arguments{
\item{accountId}{Account Id}

\item{webPropertyId}{Web Property Id}

\item{profileId}{Profile Id}
}
\value{
Experiments List
}
\description{
List Experiments
}
\seealso{
Other managementAPI functions: \code{\link{ga_experiment}},
  \code{\link{ga_filter_add}},
  \code{\link{ga_filter_apply_to_view}},
  \code{\link{ga_filter_update_filter_link}},
  \code{\link{ga_filter_update}},
  \code{\link{ga_segment_list}},
  \code{\link{ga_unsampled_list}},
  \code{\link{ga_unsampled}}
}
\concept{managementAPI functions}
