% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_management.R
\name{ga_filter}
\alias{ga_filter}
\title{Get specific filter for account}
\usage{
ga_filter(accountId, filterId)
}
\arguments{
\item{accountId}{Account Id}

\item{filterId}{Filter Id}
}
\value{
filter list
}
\description{
Get specific filter for account
}
\seealso{
Other filter management functions: \code{\link{ga_filter_delete}},
  \code{\link{ga_filter_list}},
  \code{\link{ga_filter_view_list}},
  \code{\link{ga_filter_view}}
}
\concept{filter management functions}
