% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_management.R
\name{ga_filter_apply_to_view}
\alias{ga_filter_apply_to_view}
\title{Apply an existing filter to view.}
\usage{
ga_filter_apply_to_view(filterId, accountId, webPropertyId, viewId)
}
\arguments{
\item{filterId}{The id of the filter to be added to profile/view}

\item{accountId}{Account Id of the account that contains the filter}

\item{webPropertyId}{Web property Id to create profile filter link for}

\item{viewId}{Profile/view Id to create profile filter link for}
}
\value{
A profileFilterLink object
}
\description{
Apply an existing filter to view.
}
\seealso{
Other managementAPI functions: \code{\link{ga_experiment_list}},
  \code{\link{ga_experiment}}, \code{\link{ga_filter_add}},
  \code{\link{ga_filter_update_filter_link}},
  \code{\link{ga_filter_update}},
  \code{\link{ga_segment_list}}
}
\concept{managementAPI functions}
