% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_management.R
\name{ga_filter_delete}
\alias{ga_filter_delete}
\title{Delete a filter from account or remove from view.}
\usage{
ga_filter_delete(accountId, webPropertyId = NULL, viewId = NULL,
  filterId, removeFromView = FALSE)
}
\arguments{
\item{accountId}{Account Id of the account that contains the filter}

\item{webPropertyId}{Property Id of the property that contains the filter}

\item{viewId}{View Id of the view that contains the filter}

\item{filterId}{Filter Id of the filter to be deleted}

\item{removeFromView}{Default if FALSE. If TRUE, deletes the filter from the view}
}
\value{
TRUE if successful
}
\description{
Delete a filter from account or remove from view.
}
\seealso{
Other filter management functions: \code{\link{ga_filter_list}},
  \code{\link{ga_filter_view_list}},
  \code{\link{ga_filter_view}}, \code{\link{ga_filter}}
}
\concept{filter management functions}
