% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_web_functions.R
\name{ga_filter_dsl}
\alias{ga_filter_dsl}
\title{DSL for GA4 filters}
\usage{
ga_filter_dsl(x)
}
\arguments{
\item{x}{DSL enabled syntax}
}
\description{
DSL for GA4 filters
}
\examples{

## filter clauses
# or string filter
ga_filter_dsl("city"=="Copenhagen" | "city" == "London")
# inlist string filter
ga_filter_dsl("city"==c("Copenhagen","London"))
# and string filters
ga_filter_dsl("city"=="Copenhagen" & "dayOfWeek" == "5")
# invert string filter
ga_filter_dsl(!("city"=="Copenhagen" | "city" == "London"))

## numeric filter types
# numeric equal filter
ga_filter_dsl("sessions"==5)
# between numeric filter
ga_filter_dsl("sessions"==c(5,6))
# greater than numeric
ga_filter_dsl("sessions" > 0)
# great than or equal
ga_filter_dsl("sessions" >= 1)
# less than numeric
ga_filter_dsl("sessions" < 100)
# less than or equal numeric
ga_filter_dsl("sessions" <= 100)

## string filter types
# begins with string
ga_filter_dsl("city" \%begins\% "Cope")
# ends with string
ga_filter_dsl("city" \%ends\% "hagen")
# contains string
ga_filter_dsl("city" \%in\% "ope")
# regex (full) string
ga_filter_dsl("city" \%regex\% "^Cope")
# regex (partial) string
ga_filter_dsl("city" \%regex_partial\% "ope")

# by default string filters are case sensitive.  
# Use UPPERCASE operator to make then case insensitive

# begins with string (case insensitive)
ga_filter_dsl("city" \%BEGINS\% "cope")
# ends with string (case insensitive)
ga_filter_dsl("city" \%ENDS\% "Hagen")


}
