% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_management.R
\name{ga_filter_list}
\alias{ga_filter_list}
\title{List filters for account}
\usage{
ga_filter_list(accountId)
}
\arguments{
\item{accountId}{Account Id}
}
\value{
filter list
}
\description{
List filters for account
}
\seealso{
Other filter management functions: \code{\link{ga_filter_delete}},
  \code{\link{ga_filter_view_list}},
  \code{\link{ga_filter_view}}, \code{\link{ga_filter}}
}
\concept{filter management functions}
