% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_management.R
\name{ga_filter_update_filter_link}
\alias{ga_filter_update_filter_link}
\title{Update an existing profile filter link. Patch semantics supported}
\usage{
ga_filter_update_filter_link(viewFilterLink, accountId, webPropertyId, viewId,
  linkId, method = c("PUT", "PATCH"))
}
\arguments{
\item{viewFilterLink}{The profileFilterLink object}

\item{accountId}{Account Id of the account that contains the filter}

\item{webPropertyId}{Web property Id to which the profile filter link belongs}

\item{viewId}{View Id to which the profile filter link belongs}

\item{linkId}{The id of the profile filter link to be updated}

\item{method}{PUT by default. Supports patch semantics when set to PATCH}
}
\description{
Update an existing profile filter link. Patch semantics supported
}
\examples{

\dontrun{

## Changing the rank of the filter in the view:
viewFilterLink <- list(rank = 4)

ga_filter_update_filter_link(viewFilterLink,
                             accountId, 
                             webPropertyId,
                             viewId,
                             linkId, 
                             method = "PATCH")
}

}
\seealso{
\url{https://developers.google.com/analytics/devguides/config/mgmt/v3/mgmtReference/management/profileFilterLinks}

Other managementAPI functions: \code{\link{ga_accounts}},
  \code{\link{ga_adwords_list}}, \code{\link{ga_adwords}},
  \code{\link{ga_custom_vars_list}},
  \code{\link{ga_custom_vars}},
  \code{\link{ga_experiment_list}},
  \code{\link{ga_experiment}}, \code{\link{ga_filter_add}},
  \code{\link{ga_filter_apply_to_view}},
  \code{\link{ga_filter_delete}},
  \code{\link{ga_filter_list}},
  \code{\link{ga_filter_update}},
  \code{\link{ga_filter_view_list}},
  \code{\link{ga_filter_view}}, \code{\link{ga_filter}},
  \code{\link{ga_goal_list}}, \code{\link{ga_goal}},
  \code{\link{ga_remarketing_estimate}},
  \code{\link{ga_remarketing_get}},
  \code{\link{ga_remarketing_list}},
  \code{\link{ga_segment_list}},
  \code{\link{ga_unsampled_list}},
  \code{\link{ga_unsampled}}, \code{\link{ga_users_list}},
  \code{\link{ga_view_list}}, \code{\link{ga_view}},
  \code{\link{ga_webproperty_list}},
  \code{\link{ga_webproperty}},
  \code{\link{google_analytics_account_list}}
}
