% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_management.R
\name{ga_filter_view}
\alias{ga_filter_view}
\title{Get specific filter for view (profile)}
\usage{
ga_filter_view(accountId, webPropertyId, viewId, linkId)
}
\arguments{
\item{accountId}{Account Id}

\item{webPropertyId}{Web Property Id}

\item{viewId}{Profile Id}

\item{linkId}{Link Id}
}
\value{
filter list
}
\description{
Get specific filter for view (profile)
}
\seealso{
Other filter management functions: \code{\link{ga_filter_delete}},
  \code{\link{ga_filter_list}},
  \code{\link{ga_filter_view_list}},
  \code{\link{ga_filter}}
}
\concept{filter management functions}
