% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_management.R
\name{ga_filter_view_list}
\alias{ga_filter_view_list}
\title{List filters for view (profile)}
\usage{
ga_filter_view_list(accountId, webPropertyId, viewId)
}
\arguments{
\item{accountId}{Account Id}

\item{webPropertyId}{Web Property Id}

\item{viewId}{Profile Id}
}
\value{
filter list
}
\description{
List filters for view (profile)
}
\seealso{
Other filter management functions: \code{\link{ga_filter_delete}},
  \code{\link{ga_filter_list}},
  \code{\link{ga_filter_view}}, \code{\link{ga_filter}}
}
\concept{filter management functions}
