% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/goal_management.R
\name{ga_goal}
\alias{ga_goal}
\title{Get goal}
\usage{
ga_goal(accountId, webPropertyId, profileId, goalId)
}
\arguments{
\item{accountId}{Account Id}

\item{webPropertyId}{Web Property Id}

\item{profileId}{Profile Id}

\item{goalId}{Goal Id}
}
\value{
Goal meta data
}
\description{
Get goal
}
\seealso{
Other goal management functions: 
\code{\link{ga_goal_add}()},
\code{\link{ga_goal_list}()},
\code{\link{ga_goal_update}()}
}
\concept{goal management functions}
