% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/goal_management.R
\name{ga_goal_list}
\alias{ga_goal_list}
\title{List goals}
\usage{
ga_goal_list(accountId, webPropertyId, profileId)
}
\arguments{
\item{accountId}{Account Id}

\item{webPropertyId}{Web Property Id}

\item{profileId}{Profile Id}
}
\value{
Goal list
}
\description{
List goals
}
\seealso{
Other goal management functions: \code{\link{ga_goal_add}},
  \code{\link{ga_goal_update}}, \code{\link{ga_goal}}
}
\concept{goal management functions}
