% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experiments.R, R/goals.R, R/unsampled.R
\name{ga_goals_list}
\alias{ga_goals_list}
\title{List Experiments}
\usage{
ga_goals_list(accountId, webPropertyId, profileId)

ga_goals_list(accountId, webPropertyId, profileId)

ga_goals_list(accountId, webPropertyId, profileId)
}
\arguments{
\item{accountId}{Account Id}

\item{webPropertyId}{Web Property Id}

\item{profileId}{Profile Id}

\item{accountId}{Account Id}

\item{webPropertyId}{Web Property Id}

\item{profileId}{Profile Id}

\item{accountId}{Account Id}

\item{webPropertyId}{Web Property Id}

\item{profileId}{Profile Id}
}
\value{
Experiments List

Goal list

Unsampled Reports List
}
\description{
List Experiments

List goals

List Unsampled Reports
}

