% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta.R
\name{ga_meta}
\alias{ga_meta}
\title{Get current dimensions and metrics available in GA API.}
\usage{
ga_meta(version = c("universal", "data"))
}
\arguments{
\item{version}{The Google Analytics API metadata to fetch - "universal" for Universal and earlier versions, "data" for Google Analytics 4}
}
\value{
dataframe of dimensions and metrics available to use
}
\description{
Get current dimensions and metrics available in GA API.
}
\examples{

\dontrun{

# universal analytics
ga_meta()

# Google Analytics 4 metadata from the Data API
ga_meta("data")

}
}
\seealso{
\url{https://developers.google.com/analytics/devguides/reporting/metadata/v3/reference/metadata/columns/list}, \url{https://developers.google.com/analytics/trusted-testing/analytics-data/rest/v1alpha/TopLevel/getUniversalMetadata}
}
