% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{ga_model_example}
\alias{ga_model_example}
\title{Load an example model}
\usage{
ga_model_example(name = "list")
}
\arguments{
\item{name}{name of the model - set to "list" to show available files}
}
\description{
Load an example model
}
\examples{

# example .gamr files included with the package
ga_model_example()

# load one example
ga_model_example("ga4-trend.gamr")
}
\seealso{
Other GA modelling functions: 
\code{\link{ga_model_edit}()},
\code{\link{ga_model_load}()},
\code{\link{ga_model_make}()},
\code{\link{ga_model_shiny_load}()},
\code{\link{ga_model_shiny_template}()},
\code{\link{ga_model_shiny}()},
\code{\link{ga_model_write}()},
\code{\link{ga_model}()}
}
\concept{GA modelling functions}
