% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/platform.R
\name{ga_model_make}
\alias{ga_model_make}
\title{Modelling function factory for Google Analytics data}
\usage{
ga_model_make(data_f, required_columns, model_f,
  required_packages = NULL, description = NULL,
  ui_out_func = shiny::textOutput)
}
\arguments{
\item{data_f}{A function that gets the data - 
must have viewId as first argument}

\item{required_columns}{What dimensions and metrics are required}

\item{model_f}{A function that inputs gadata, and outputs a list of assets - 
must take data from result of data_f in first argument.}

\item{required_packages}{The packages needed for data_f and model_f to work}

\item{description}{An optional description of what the model does}

\item{ui_out_func}{A shiny UI output function that will display the results of model_f

The passed functions should all have \code{...} to make them flexible 
  in what arguments can be added.  Do not have the same argument names in both functions.}
}
\description{
Modelling function factory for Google Analytics data
}
\examples{

\dontrun{

 get_model_data <- function(viewId, 
                           date_range = c(Sys.Date()- 300, Sys.Date()), 
                           ...){
   google_analytics(viewId, 
                    date_range = date_range, 
                    metrics = "sessions", 
                    dimensions = "date", 
                    max = -1)
 }

 decompose_sessions <- function(df, ...){
   web_data_ts <- ts(df$sessions, frequency = 7)
   d <- decompose(web_data_ts)
   list(decom = d, plot = plot(d))
 }
 
 decomp_ga <- ga_model_make(get_model_data,
                            required_columns = c("date", "sessions"),
                            model_f = decompose_sessions,
                            description = "Performs decomposition on session 
                                           data and created a plot")
 
 # fetches data and outputs decomposition
 ga_model(81416156, decomp_ga)
 
 # save the model for later
 model_location <- "inst/models/decomp_ga.gamodel"
 ga_model_save(decomp_ga, filename = model_location)
 
 # can load model from file
 ga_model(81416156, model_location)
 
 # load model and use again
 model2 <- ga_model_load(model_location)
 
 ga_model(81416156, model2)
 
}
}
\seealso{
Other GA modelling functions: \code{\link{ga_model_load}},
  \code{\link{ga_model_save}}, \code{\link{ga_model}}
}
\concept{GA modelling functions}
