% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{ga_model_refresh}
\alias{ga_model_refresh}
\title{Refresh a model}
\usage{
ga_model_refresh(model)
}
\arguments{
\item{model}{Model or file location of model .gamr file}
}
\description{
Sometimes necessary if functions were created under differing package versions
}
\examples{

\dontrun{

decomp_ga <- ga_model_example("decomp_ga.gamr")
decomp_ga <- ga_model_refresh(decomp_ga)

}
}
