% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{ga_model_save}
\alias{ga_model_save}
\title{Save a created model}
\usage{
ga_model_save(model, filename = "my-model.gamr")
}
\arguments{
\item{model}{model to save}

\item{filename}{name to save model under}
}
\description{
Save a created model
}
\concept{GA modelling functions

\dontrun{
# load the model (equivalent to ga_model_load())
decomp_ga <- ga_model_example("decomp_ga.gamr")

# save it somewhere else
ga_model_save(decomp_ga, "somewhereelse.gamr")

}}
