% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_shiny_templates.R
\name{ga_model_shiny_ui}
\alias{ga_model_shiny_ui}
\alias{ga_model_shiny_load}
\title{How to generate the model objects HTML}
\usage{
ga_model_shiny_ui(model_n)

ga_model_shiny_load(model_n, ...)
}
\arguments{
\item{model_n}{The templated name of a model e.g. 'model1'}

\item{...}{Other arguments passed from shiny server}
}
\description{
This is the default way to create model UIs generated from the model name e.g. 'model1', 'model2' etc.
}
\details{
\code{ga_model_shiny_ui()} generates the UI for loading the model.  At minimum it needs to load \code{model1$ui('model1')} which is the default.  You may want to add some logic to make menu items for some templates, which this function helps facilitate.
}
\examples{

ga_model_shiny_ui("model1")
}
\seealso{
Other GA modelling functions: 
\code{\link{ga_model_edit}()},
\code{\link{ga_model_example}()},
\code{\link{ga_model_load}()},
\code{\link{ga_model_make}()},
\code{\link{ga_model_save}()},
\code{\link{ga_model_shiny_template}()},
\code{\link{ga_model_shiny}()},
\code{\link{ga_model_write}()},
\code{\link{ga_model}()}
}
\concept{GA modelling functions}
