% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_shiny_templates.R
\name{ga_model_shiny_template}
\alias{ga_model_shiny_template}
\title{Get a Shiny template file}
\usage{
ga_model_shiny_template(name = "list", read_lines = FALSE)
}
\arguments{
\item{name}{the template name}

\item{read_lines}{If TRUE will use \code{readLines()} to print out the template contents}
}
\description{
Gets a pre-created template from the googleAnalyticsR samples
}
\seealso{
Other GA modelling functions: 
\code{\link{ga_model_edit}()},
\code{\link{ga_model_example}()},
\code{\link{ga_model_load}()},
\code{\link{ga_model_make}()},
\code{\link{ga_model_shiny_load}()},
\code{\link{ga_model_shiny}()},
\code{\link{ga_model_write}()},
\code{\link{ga_model}()}
}
\concept{GA modelling functions}
