% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/platform.R
\name{ga_model_tweet}
\alias{ga_model_tweet}
\title{Upload an interactive visualisation so it can be embedded in a tweet}
\usage{
ga_model_tweet(model_output, twitter, title, bucket, image = "")
}
\arguments{
\item{model_output}{A \code{ga_model_result} object created by \link{ga_model}}

\item{twitter}{Your twitter handle e.g. \code{@holomarked}}

\item{title}{Twitter preview card title text}

\item{bucket}{The GCS bucket to upload to}

\item{image}{an optional image to display before the visualition runs}
}
\description{
Inspired by \url{https://datatitian.com/how-to-turn-your-ggplot2-visualization-into-an-interactive-tweet/} this uploads your model output into Google Cloud storage, in the right format to embd in a tweet
}
\details{
Authenticate with `googleCloudStorageR` first.  \url{https://cards-dev.twitter.com/validator} is useful to test what it will look like on Twitter.
}
\examples{

\dontrun{
  library(googleAnalyticsModelR)
  library(googleAnalyticsR)

  output <- ga_time_normalised(81416156, interactive_plot = TRUE)
  
  # if you have a plotly account, get a static image
  plotly_IMAGE(output$plot, out_file = "tweet.png")

  library(googleCloudStorageR)
  ga_model_tweet(output, 
                 "@HoloMarked", 
                 "Test2 ga_model twitter upload", 
                 bucket = "mark-edmondson-public-read",
                 image = "tweet.png")
}

}
