% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measurement_protocol.R
\name{ga_mp_cid}
\alias{ga_mp_cid}
\title{Generate a random client_id}
\usage{
ga_mp_cid(seed = NULL)
}
\arguments{
\item{seed}{If you set a seed, then the random number will be the same for each value}
}
\description{
This has a random number plus a timestamp
}
\seealso{
Other Measurement Protocol functions: 
\code{\link{ga_mp_event_item}()},
\code{\link{ga_mp_event}()},
\code{\link{ga_mp_send}()}
}
\concept{Measurement Protocol functions}
