% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measurement_protocol.R
\name{ga_mp_event}
\alias{ga_mp_event}
\title{Create a Measurement Protocol Event}
\usage{
ga_mp_event(name, params = NULL, items = NULL)
}
\arguments{
\item{name}{The event name to send in}

\item{items}{Optional items created via \link{ga_mp_event_item}}

\item{param}{Optional event parameters sent in as a named list}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
This creates an event to send via \link{ga_mp_send}
}
\examples{

ga_mp_event("custom_event")
ga_mp_event("custom_event", params = list(my_param = "SUPER"))
}
\seealso{
Other Measurement Protocol functions: 
\code{\link{ga_mp_event_item}()},
\code{\link{ga_mp_send}()}
}
\concept{Measurement Protocol functions}
