% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measurement_protocol.R
\name{ga_mp_send}
\alias{ga_mp_send}
\title{Make a Measurement Protocol v2 request}
\usage{
ga_mp_send(
  events,
  measurement_id,
  client_id,
  user_id = NULL,
  api_secret = Sys.getenv("GA_MP_SECRET"),
  debug = FALSE,
  timestamp_micros = NULL,
  user_properties = NULL,
  non_personalized_ads = TRUE
)
}
\arguments{
\item{events}{The events to send}

\item{measurement_id}{The measurement ID associated with a stream}

\item{client_id}{The client_id to associate with the event}

\item{user_id}{Optional. Unique id for the user}

\item{api_secret}{The secret generated in the GA4 UI - by default will look for environment arg \code{GA_MP_SECRET}}

\item{debug}{Send hits to the debug endpoint to validate hits}

\item{timestamp_micros}{Optional. A Unix timestamp (in microseconds) for the time to associate with the event.}

\item{user_properties}{Optional. The user properties for the measurement.}

\item{non_personalized_ads}{Optional. Set to true to indicate these events should not be used for personalized ads.}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
Create a server side call to Google Analytics 4 via its Measurement Protocol
}
\details{
Create an API secret via \verb{Admin > Data Streams > choose your stream > Measurement Protocol > Create}

\code{user_id} can be used for \href{https://support.google.com/analytics/answer/9213390}{cross-platform analysis}

\code{timestamp_micros} should only be set to record events that happened in the past. This value can be overridden via user_property or event timestamps. Events can be backdated up to 48 hours. Note microseconds, not milliseconds.

\code{user_properties} - describe segments of your user base, such as language preference or geographic location.  See \href{https://developers.google.com/analytics/devguides/collection/protocol/ga4/user-properties?client_type=gtag}{User properties}
}
\section{ClientID}{


In order for an event to be valid, it must have a client_id that has already been used to send an event from gtag.js. You will need to capture this ID client-side and include it in your call to the measurement protocol.  Ensure you also have user permission as specified in the \href{https://developers.google.com/analytics/devguides/collection/protocol/ga4/policy}{feature policy}
}

\examples{
# preferably set this in .Renviron
Sys.setenv(GA_MP_SECRET="MY_SECRET")

# your GA4 settings
my_measurement_id <- "G-43MDXK6CLZ"
a_client_id <- 1234567
S

event <- ga_mp_event("an_event")
ga_mp_send(event, 1234567, "G-43MDXK6CLZ", debug = TRUE)

another <- ga_mp_event("another_event")
ga_mp_send(list(event, another), 1234567, "G-43MDXK6CLZ", debug = TRUE)

}
\seealso{
\href{https://developers.google.com/analytics/devguides/collection/protocol/ga4}{Measurement Protocol (Google Analytics 4)}

Other Measurement Protocol functions: 
\code{\link{ga_mp_event_item}()},
\code{\link{ga_mp_event}()}
}
\concept{Measurement Protocol functions}
