% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remarketingAudiences.R
\name{ga_remarketing_estimate}
\alias{ga_remarketing_estimate}
\title{Estimate number of users added to the segment yesterday}
\usage{
ga_remarketing_estimate(remarketingAudience)
}
\arguments{
\item{remarketingAudience}{A remarketing audience object from \link{ga_remarketing_get}

Takes the segment definition from a remarketing audiences and runs 
  it against the viewId to see current estimated users

The total audience size is this figure for every membershipDurationDay from yesterday}
}
\value{
data.frame
}
\description{
Estimate number of users added to the segment yesterday
}
\seealso{
\href{https://support.google.com/analytics/answer/2611268?hl=en}{About remarketing audiences}

Other remarketing management functions: \code{\link{ga_remarketing_build}},
  \code{\link{ga_remarketing_create}},
  \code{\link{ga_remarketing_get}},
  \code{\link{ga_remarketing_list}}
}
\concept{remarketing management functions}
