% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remarketingAudiences.R
\name{ga_remarketing_get}
\alias{ga_remarketing_get}
\title{Get a remarketing audience}
\usage{
ga_remarketing_get(accountId, webPropertyId, remarketingAudienceId)
}
\arguments{
\item{accountId}{Account Id}

\item{webPropertyId}{Web Property Id}

\item{remarketingAudienceId}{The ID of the remarketing audience to retrieve.}
}
\value{
Remarketing Audience object
}
\description{
Get a remarketing audience
}
\seealso{
\href{https://support.google.com/analytics/answer/2611268?hl=en}{About remarketing audiences}

Other remarketing management functions: 
\code{\link{ga_remarketing_build}()},
\code{\link{ga_remarketing_create}()},
\code{\link{ga_remarketing_estimate}()},
\code{\link{ga_remarketing_list}()}
}
\concept{remarketing management functions}
