% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remarketingAudiences.R
\name{ga_remarketing_list}
\alias{ga_remarketing_list}
\title{List remarketing audiences}
\usage{
ga_remarketing_list(accountId, webPropertyId)
}
\arguments{
\item{accountId}{Account Id}

\item{webPropertyId}{Web Property Id}
}
\value{
Remarketing audience list
}
\description{
List remarketing audiences
}
\seealso{
\href{https://support.google.com/analytics/answer/2611268?hl=en}{About remarketing audiences}

Other remarketing management functions: 
\code{\link{ga_remarketing_build}()},
\code{\link{ga_remarketing_create}()},
\code{\link{ga_remarketing_estimate}()},
\code{\link{ga_remarketing_get}()}
}
\concept{remarketing management functions}
