% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segments.R
\name{ga_segment_list}
\alias{ga_segment_list}
\title{Get segments user has access to}
\usage{
ga_segment_list()
}
\value{
Segment list
}
\description{
Get segments user has access to
}
\seealso{
Other managementAPI functions: \code{\link{ga_experiment_list}},
  \code{\link{ga_experiment}}, \code{\link{ga_filter_add}},
  \code{\link{ga_filter_apply_to_view}},
  \code{\link{ga_filter_update_filter_link}},
  \code{\link{ga_filter_update}}
}
\concept{managementAPI functions}
