% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measurement_protocol.R
\name{ga_trackme}
\alias{ga_trackme}
\title{Opt in or out of googleAnalyticsR usage tracking}
\usage{
ga_trackme()
}
\description{
You can opt-in or out to sending a measurement protocol hit when you load the package for use in the package's statistics via this function, or opt-out again by deleting the \code{.optin-googleanalyticsr} file.
}
\details{
Run this function to control your choices.
}
\seealso{
Other package tracking functions: 
\code{\link{ga_trackme_event}()}
}
\concept{package tracking functions}
