% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unsampled.R
\name{ga_unsampled}
\alias{ga_unsampled}
\title{Get Unsampled Report Meta Data}
\usage{
ga_unsampled(accountId, webPropertyId, profileId, unsampledReportId)
}
\arguments{
\item{accountId}{Account Id}

\item{webPropertyId}{Web Property Id}

\item{profileId}{Profile Id}

\item{unsampledReportId}{Unsampled Report Id}
}
\value{
Unsampled Report Meta Data
}
\description{
Get Unsampled Report Meta Data
}
\seealso{
Other managementAPI functions: \code{\link{ga_experiment_list}},
  \code{\link{ga_experiment}}, \code{\link{ga_filter_add}},
  \code{\link{ga_filter_apply_to_view}},
  \code{\link{ga_filter_update_filter_link}},
  \code{\link{ga_filter_update}},
  \code{\link{ga_segment_list}},
  \code{\link{ga_unsampled_list}}
}
\concept{managementAPI functions}
