% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unsampled.R
\name{ga_unsampled_download}
\alias{ga_unsampled_download}
\title{Download Unsampled Report from Google Drive}
\usage{
ga_unsampled_download(reportTitle, accountId, webPropertyId, profileId,
  downloadFile = TRUE)
}
\arguments{
\item{reportTitle}{Title of Unsampled Report (case-sensitive)}

\item{accountId}{Account Id}

\item{webPropertyId}{Web Property Id}

\item{profileId}{Profile Id}

\item{downloadFile}{Default TRUE, whether to download, if FALSE returns a dataframe instead}
}
\value{
file location if \code{downloadFile} is TRUE, else a \code{data.frame} of download
}
\description{
Download Unsampled Report from Google Drive
}
