% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filters.R
\name{ga_update_filter_patch}
\alias{ga_update_filter_patch}
\title{Updates an existing filter. This method supports patch semantics.}
\usage{
ga_update_filter_patch(Filter, accountId, filterId)
}
\arguments{
\item{Filter}{The list to be turned to JSON to create a request body}

\item{accountId}{Account Id of the account that contains the filter}

\item{filterId}{The id of the filter to be modified}
}
\description{
Updates an existing filter. This method supports patch semantics.
}
\seealso{
Other managementAPI functions: \code{\link{ga_accounts}},
  \code{\link{ga_add_filter}},
  \code{\link{ga_add_profile_filter}},
  \code{\link{ga_adwords_list}}, \code{\link{ga_adwords}},
  \code{\link{ga_custom_vars_list}},
  \code{\link{ga_custom_vars}},
  \code{\link{ga_delete_filter}},
  \code{\link{ga_delete_profile_filter}},
  \code{\link{ga_experiment_list}},
  \code{\link{ga_experiment}},
  \code{\link{ga_filter_list}},
  \code{\link{ga_filter_view_list}},
  \code{\link{ga_filter_view}}, \code{\link{ga_filter}},
  \code{\link{ga_goal_list}}, \code{\link{ga_goal}},
  \code{\link{ga_remarketing_estimate}},
  \code{\link{ga_remarketing_get}},
  \code{\link{ga_remarketing_list}},
  \code{\link{ga_segment_list}},
  \code{\link{ga_unsampled_list}},
  \code{\link{ga_unsampled}},
  \code{\link{ga_update_filter}},
  \code{\link{ga_update_profile_filter_patch}},
  \code{\link{ga_update_profile_filter}},
  \code{\link{ga_users_list}}, \code{\link{ga_view_list}},
  \code{\link{ga_view}}, \code{\link{ga_webproperty_list}},
  \code{\link{ga_webproperty}},
  \code{\link{google_analytics_account_list}}
}
