% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filters.R
\name{ga_update_profile_filter}
\alias{ga_update_profile_filter}
\title{Update an existing profile filter link.}
\usage{
ga_update_profile_filter(profileFilterLink, accountId, webPropertyId, profileId,
  linkId)
}
\arguments{
\item{profileFilterLink}{The list to be turned to JSON to create a request body}

\item{accountId}{Account Id of the account that contains the filter}

\item{webPropertyId}{Web property Id to which the profile filter link belongs}

\item{profileId}{Profile/view Id to which the profile filter link belongs}

\item{linkId}{The id of the profile filter link to be updated}
}
\description{
Update an existing profile filter link.
}
\seealso{
Other managementAPI functions: \code{\link{ga_accounts}},
  \code{\link{ga_add_filter}},
  \code{\link{ga_add_profile_filter}},
  \code{\link{ga_adwords_list}}, \code{\link{ga_adwords}},
  \code{\link{ga_custom_vars_list}},
  \code{\link{ga_custom_vars}},
  \code{\link{ga_delete_filter}},
  \code{\link{ga_delete_profile_filter}},
  \code{\link{ga_experiment_list}},
  \code{\link{ga_experiment}},
  \code{\link{ga_filter_list}},
  \code{\link{ga_filter_view_list}},
  \code{\link{ga_filter_view}}, \code{\link{ga_filter}},
  \code{\link{ga_goal_list}}, \code{\link{ga_goal}},
  \code{\link{ga_remarketing_estimate}},
  \code{\link{ga_remarketing_get}},
  \code{\link{ga_remarketing_list}},
  \code{\link{ga_segment_list}},
  \code{\link{ga_unsampled_list}},
  \code{\link{ga_unsampled}},
  \code{\link{ga_update_filter_patch}},
  \code{\link{ga_update_filter}},
  \code{\link{ga_update_profile_filter_patch}},
  \code{\link{ga_users_list}}, \code{\link{ga_view_list}},
  \code{\link{ga_view}}, \code{\link{ga_webproperty_list}},
  \code{\link{ga_webproperty}},
  \code{\link{google_analytics_account_list}}
}
