% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/views.R
\name{ga_view}
\alias{ga_view}
\title{Get single View (Profile)}
\usage{
ga_view(accountId, webPropertyId, profileId)
}
\arguments{
\item{accountId}{Account Id}

\item{webPropertyId}{Web Property Id}

\item{profileId}{Profile (View) Id}
}
\value{
A list of the Views meta-data.
}
\description{
Gets meta-data for a particular View/Profile
}
\examples{

\dontrun{

library(googleAnalyticsR)
ga_auth()
ga_view(1058095, webPropertyId = "UA-1058095-1", profileId = 1855267)
}
}
\seealso{
Other account structure functions: \code{\link{ga_accounts}},
  \code{\link{ga_view_list}},
  \code{\link{ga_webproperty_list}},
  \code{\link{ga_webproperty}},
  \code{\link{google_analytics_account_list}}
}
\concept{account structure functions}
