% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/webprops.R
\name{ga_webproperty}
\alias{ga_webproperty}
\title{Get web property}
\usage{
ga_webproperty(accountId, webPropertyId)
}
\arguments{
\item{accountId}{Account Id}

\item{webPropertyId}{Web Property Id}
}
\value{
webproperty
}
\description{
Get web property
}

