% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_web_functions.R
\name{gaw_filter_expression}
\alias{gaw_filter_expression}
\title{Create a filter for use with App+Web Reports}
\usage{
gaw_filter_expression(a_filter, type = c("default", "not", "and", "or"))
}
\arguments{
\item{a_filter}{A Filter object or a FilterExpression or a list of FilterExpressions}

\item{type}{The type of filter}
}
\description{
For use with \link{google_analytics_aw}
}
\details{
The fields in the same FilterExpression need to be either all dimensions or all metrics.
}
