% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ga_v4_get.R
\name{google_analytics_4}
\alias{google_analytics_4}
\title{Do a single request}
\usage{
google_analytics_4(viewId, date_range, metrics, dimensions = NULL,
  date_range_two = NULL, dim_filters = NULL, met_filters = NULL,
  filtersExpression = NULL, segments = NULL, pivots = NULL, max = 1000,
  samplingLevel = c("DEFAULT", "SMALL", "LARGE"), metricFormat = NULL,
  histogramBuckets = NULL)
}
\arguments{
\item{viewId}{viewId of data to get.}

\item{date_range}{character or date vector of format c(start, end). Optionc("#752324", "white", "white", "indianred")c("#FFFFFF")al c(s1,e1,s2,e2)}

\item{metrics}{Metric to fetch. Required. Supports calculated metrics.}

\item{dimensions}{Dimensions to fetch.}

\item{segments}{Segments of the data}

\item{pivots}{Pivots of the data}

\item{samplingLevel}{Sample level}

\item{metricFormat}{If supplying calculated metrics, specify the metric type}
}
\description{
Do a single request
}

