% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_functions.R
\name{google_analytics_4_parse}
\alias{google_analytics_4_parse}
\title{ga v4 data parsing}
\usage{
google_analytics_4_parse(x)
}
\description{
x is response_list$reports[[1]] from google_analytics_4_parse_batch
}
\keyword{internal}
