% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_functions.R
\name{google_analytics_4_parse_batch}
\alias{google_analytics_4_parse_batch}
\title{ga v4 parse batching}
\usage{
google_analytics_4_parse_batch(response_list)
}
\description{
ga v4 parse batching
}
\keyword{internal}
