% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_functions.R
\name{google_analytics_4_parse_tidyjson}
\alias{google_analytics_4_parse_tidyjson}
\title{seeing if this is better}
\usage{
google_analytics_4_parse_tidyjson(x)
}
\description{
seeing if this is better
}

