% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bigQuery.R
\name{google_analytics_bq_asynch}
\alias{google_analytics_bq_asynch}
\title{Asynch fetch}
\usage{
google_analytics_bq_asynch(projectId, datasetId, query, bucket, download_file)
}
\description{
Asynch fetch
}
\keyword{internal}
