% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ga4_data_objects_filters.R
\name{InListFilter}
\alias{InListFilter}
\title{InListFilter Object}
\usage{
InListFilter(values, caseSensitive = TRUE)
}
\arguments{
\item{values}{The vector of string values}

\item{caseSensitive}{If true, the string value is case sensitive}
}
\value{
InListFilter object
}
\description{
InListFilter Object
}
\details{
The result needs to be in a list of string values.
}
\concept{InListFilter functions}
