% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{is.error}
\alias{is.error}
\title{Is this a try error?}
\usage{
is.error(test_me)
}
\arguments{
\item{test_me}{an object created with try()}
}
\value{
Boolean
}
\description{
Utility to test errors
}
\keyword{internal}
