% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{listNameToDFCol}
\alias{listNameToDFCol}
\title{Add name of list entry of dataframe to dataframe colum}
\usage{
listNameToDFCol(named_list, colName = "listName")
}
\description{
Add name of list entry of dataframe to dataframe colum
}
\keyword{internal}
