% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ga_v4_get.R
\name{make_ga_4_req}
\alias{make_ga_4_req}
\title{Google Analytics v4 API fetch}
\usage{
make_ga_4_req(viewId, date_range = NULL, metrics = NULL,
  dimensions = NULL, dim_filters = NULL, met_filters = NULL,
  filtersExpression = NULL, orderBys = NULL, segments = NULL,
  pivots = NULL, cohorts = NULL, pageToken = 1, pageSize = 1000,
  samplingLevel = c("DEFAULT", "SMALL", "LARGE"), metricFormat = NULL,
  histogramBuckets = NULL)
}
\arguments{
\item{viewId}{viewId of data to get.}

\item{date_range}{character or date vector of format c(start, end). Optional c(s1,e1,s2,e2)}

\item{metrics}{Metric to fetch. Required. Supports calculated metrics.}

\item{dimensions}{Dimensions to fetch.}

\item{segments}{Segments of the data}

\item{pivots}{Pivots of the data}

\item{cohorts}{Cohorts created by \link{makeCohortGroup}}

\item{pageToken}{Where to start the data fetch,}

\item{pageSize}{How many rows to fetch. Max 10000 each batch.}

\item{samplingLevel}{Sample level}

\item{metricFormat}{If supplying calculated metrics, specify the metric type}

\item{filters}{A list of dimensionFilterClauses or metricFilterClauses}
}
\description{
This function constructs the Google Analytics API v4 call to be called
  via \code{fetch_google_analytics_4}
}
\section{Metrics}{

  Metrics support calculated metrics like ga:users / ga:sessions if you supply
  them in a named vector.

  You must supply the correct 'ga:' prefix unlike normal metrics

  You can mix calculated and normal metrics like so:

  \code{customMetric <- c(sessionPerVisitor = "ga:sessions / ga:visitors",
                          "bounceRate",
                          "entrances")}

   You can also optionally supply a \code{metricFormat} parameter that must be
   the same length as the metrics.  \code{metricFormat} can be:
   \code{METRIC_TYPE_UNSPECIFIED, INTEGER, FLOAT, CURRENCY, PERCENT, TIME}

   All metrics are currently parsed to as.numeric when in R.
}

\section{Dimensions}{


  Supply a character vector of dimensions, with or without \code{ga:} prefix.

  Optionally for numeric dimension types such as
  \code{ga:hour, ga:browserVersion, ga:sessionsToTransaction}, etc. supply
  histogram buckets suitable for histogram plots.

  If non-empty, we place dimension values into buckets after string to int64.
  Dimension values that are not the string representation of an integral value
  will be converted to zero. The bucket values have to be in increasing order.
  Each bucket is closed on the lower end, and open on the upper end.
  The "first" bucket includes all values less than the first boundary,
  the "last" bucket includes all values up to infinity.
  Dimension values that fall in a bucket get transformed to a new dimension
  value. For example, if one gives a list of "0, 1, 3, 4, 7", then we
  return the following buckets: -

  bucket #1: values < 0, dimension value "<0"
  bucket #2: values in [0,1), dimension value "0"
  bucket #3: values in [1,3), dimension value "1-2"
  bucket #4: values in [3,4), dimension value "3"
  bucket #5: values in [4,7), dimension value "4-6"
  bucket #6: values >= 7, dimension value "7+"
}

