% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pivots.R
\name{makePivotData}
\alias{makePivotData}
\title{Make pivot data rows}
\usage{
makePivotData(pivotRows, pivotNames, index = 1)
}
\arguments{
\item{pivotRows}{from x$data$rows}

\item{index}{1 for normal data, 2 for date comparison}
}
\description{
Make pivot data rows
}
\seealso{
Other pivot functions: \code{\link{makePivotNames}},
  \code{\link{pivot_ga4_parse}}
}
\keyword{internal}
