% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pivots.R
\name{makePivotNames}
\alias{makePivotNames}
\title{Make pivot column names}
\usage{
makePivotNames(pivotHeaders)
}
\description{
Make pivot column names
}
\seealso{
Other pivot functions: \code{\link{makePivotData}},
  \code{\link{pivot_ga4_parse}}
}
\keyword{internal}
