% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filters.R
\name{met_filter}
\alias{met_filter}
\title{Make a metric filter object}
\usage{
met_filter(metric, operator = c("EQUAL", "LESS_THAN", "GREATER_THAN",
  "IS_MISSING"), comparisonValue, not = FALSE)
}
\arguments{
\item{metric}{metric name to filter on.}

\item{operator}{How to match the dimension.}

\item{comparisonValue}{What to match.}

\item{not}{Logical NOT operator. Boolean.}
}
\value{
An object of class \code{met_fil_ga4} for use in \code{\link{filter_clause_ga4}}
}
\description{
Make a metric filter object
}
\seealso{
Other filter functions: \code{\link{dim_filter}},
  \code{\link{filter_clause_ga4}}
}

