% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{meta}
\alias{meta}
\title{Google Analytics API metadata}
\format{A data frame with 476 rows and 15 variables}
\source{
\url{https://developers.google.com/analytics/devguides/reporting/core/dimsmets}
}
\usage{
meta
}
\description{
This is a local copy of the data provided by \link{google_analytics_meta}
}
\details{
Running your own call will be more up to date, but this is here in case.

It does not include the multi-channel or cohort variables.
}
\keyword{datasets}
