% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{meta4}
\alias{meta4}
\title{Google Analytics API metadata}
\format{
A data frame containing metric and dimensions that 
  you can query the Data API with.
}
\source{
\url{https://developers.google.com/analytics/devguides/reporting/data/v1/api-schema}
}
\usage{
meta4
}
\description{
This is a local copy of the data provided by \link{ga_meta("data")}
}
\details{
Running your own call will be more up to date, but this is here in case.
}
\keyword{datasets}
