% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_web.R
\name{Metric}
\alias{Metric}
\title{Metric Object}
\usage{
Metric(expression = NULL, invisible = NULL, name = NULL)
}
\arguments{
\item{expression}{A mathematical expression for derived metrics}

\item{invisible}{Indicates if a metric is invisible}

\item{name}{The name of the metric}
}
\value{
Metric object
}
\description{
Metric Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The quantitative measurements of a report. For example, the metric eventCount is the total number of events. Requests are allowed up to 10 metrics.
}
\concept{Metric functions}
