% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ga_v4_objects.R
\name{metric_ga4}
\alias{metric_ga4}
\title{Make a metric object}
\usage{
metric_ga4(vector, metricFormat = NULL)
}
\arguments{
\item{vector}{A character vector of GA metric names}

\item{metricFormat}{Optional same length character vector of formats
for the API to return}
}
\value{
a list suitable for parsing in req
}
\description{
Make a metric object
}
\keyword{internal}
