% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny-modules.R
\name{multi_select}
\alias{multi_select}
\title{multi_select}
\usage{
multi_select(input, output, session, type = c("METRIC", "DIMENSION"),
  subType = c("none", "segment", "cohort"))
}
\arguments{
\item{input}{shiny input}

\item{output}{shiny output}

\item{session}{shiny session}

\item{type}{metric or dimension}
}
\value{
the selected variable
}
\description{
Shiny Module for use with \link{multi_selectUI}
}
\details{
Call via \code{shiny::callModule(multi_select, "your_id")}
}

