% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny-modules.R
\name{multi_selectUI}
\alias{multi_selectUI}
\title{multi_select UI [Shiny Module]}
\usage{
multi_selectUI(id, label = "Metric", multiple = TRUE, width = NULL)
}
\arguments{
\item{id}{Shiny id}

\item{label}{label}

\item{multiple}{multiple select}

\item{width}{width of select}
}
\value{
Shiny UI
}
\description{
Shiny Module for use with \link{multi_select}
}
\details{
Create a Google Analytics variable selector
}
\seealso{
Other Shiny modules: \code{\link{authDropdownUI}},
  \code{\link{authDropdown}}, \code{\link{multi_select}}
}
