% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ga4_data_objects_filters.R
\name{NumericFilter}
\alias{NumericFilter}
\title{NumericFilter Object}
\usage{
NumericFilter(
  value,
  operation = c("EQUAL", "LESS_THAN", "LESS_THAN_OR_EQUAL", "GREATER_THAN",
    "GREATER_THAN_OR_EQUAL")
)
}
\arguments{
\item{value}{A numeric value or a date value}

\item{operation}{The operation type for this filter}
}
\value{
NumericFilter object
}
\description{
NumericFilter Object
}
\details{
Filters for numeric or date values.
}
\concept{NumericFilter functions}
