% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ga_v4_objects.R
\name{orFiltersForSegment}
\alias{orFiltersForSegment}
\title{orFiltersForSegment}
\usage{
orFiltersForSegment(segmentFilterClause)
}
\description{
orFiltersForSegment
}
\seealso{
Other v4 segment functions: \code{\link{dynamicSegment}},
  \code{\link{segmentDefinition}},
  \code{\link{segmentDimensionFilter}},
  \code{\link{segmentFilterClause}},
  \code{\link{segmentFilter}},
  \code{\link{segmentMetricFilter}},
  \code{\link{segmentObj_ga4}},
  \code{\link{segmentSequenceStep}},
  \code{\link{segment_filter}}, \code{\link{segment_ga4}},
  \code{\link{sequenceSegment}},
  \code{\link{simpleSegment}}
}

