% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_functions.R
\name{parse_ga_account_summary}
\alias{parse_ga_account_summary}
\title{New parse GA account summary}
\usage{
parse_ga_account_summary(x)
}
\arguments{
\item{x}{The account summary items}
}
\description{
New parse GA account summary
}
\keyword{internal}
