% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ga_data.R
\name{parse_realtime}
\alias{parse_realtime}
\title{ga_batch_report <- function(requestObj){
  
  url <- sprintf("https://analyticsdata.googleapis.com/%s:batchRunReports",
                 version_aw())
  
  # analyticsdata.batchRunReports
  f <- gar_api_generator(url, "POST", 
                         data_parse_function = parse_batchrunreports)
  
  stopifnot(inherits(requestObj, "gar_BatchRunReportsRequest"))
  o <- f(the_body = requestObj)
  
  o
}}
\usage{
parse_realtime(x)
}
\description{
ga_batch_report <- function(requestObj){
  
  url <- sprintf("https://analyticsdata.googleapis.com/%s:batchRunReports",
                 version_aw())
  
  # analyticsdata.batchRunReports
  f <- gar_api_generator(url, "POST", 
                         data_parse_function = parse_batchrunreports)
  
  stopifnot(inherits(requestObj, "gar_BatchRunReportsRequest"))
  o <- f(the_body = requestObj)
  
  o
}
}
