% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pivots.R
\name{pivot_ga4}
\alias{pivot_ga4}
\title{Make a pivot object}
\usage{
pivot_ga4(pivot_dim, metrics, dim_filter_clause = NULL, startGroup = 1,
  maxGroupCount = 5)
}
\arguments{
\item{pivot_dim}{A character vector of dimensions}

\item{metrics}{Metrics to aggregate and return.}

\item{dim_filter_clause}{Only data included in filter included.}

\item{startGroup}{which groups of k columns are included in response.}

\item{maxGroupCount}{Maximum number of groups to return.}
}
\value{
pivot object of class \code{pivot_ga4} for use in \code{\link{filter_clause_ga4}}
}
\description{
Make a pivot object
}
\details{
If maxGroupCount is set to -1 returns all groups.
}

