% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ga_data_pivot.R
\name{PivotObject}
\alias{PivotObject}
\title{Pivot Object}
\usage{
PivotObject(
  fieldNames,
  orderBys = NULL,
  offset = NULL,
  limit = NULL,
  metricAggregations = NULL
)
}
\arguments{
\item{fieldNames}{Dimension names for visible columns in the report response}

\item{orderBys}{Specifies how dimensions are ordered in the pivot}

\item{offset}{The row count of the start row. The first row is counted as row 0}

\item{limit}{The number of rows to return in this pivot. If unspecified, 10 rows are returned. If -1, all rows are returned.}

\item{metricAggregations}{Aggregate the metrics by dimensions in this pivot using the specified metricAggregations.}
}
\value{
Pivot Object
}
\description{
Pivot Object
}
