% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ga_data_pivot.R
\name{PivotOrderBy}
\alias{PivotOrderBy}
\title{Create a GA4 Pivot object for the Data API}
\usage{
PivotOrderBy(pivotSelections = NULL, metricName = NULL)
}
\arguments{
\item{pivotSelections}{Used to select a dimension name and value pivot}

\item{metricName}{In the response to order by, order rows by this column}
}
\value{
PivotOrderBy object
}
\description{
PivotOrderBy Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Sorts by a pivot column group.
}
\concept{PivotOrderBy functions}
