% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print-method.R
\name{print.ga_custom_data_source_upload}
\alias{print.ga_custom_data_source_upload}
\title{#' @export
print.ga_profileFilterLink <- function(x, ...){
  cat("==Google Analytics Profile (View) Filter Link==\n")
  cat0("profileFilterLink ID:  ", x$id)
  cat0("Rank:                  ", x$rank)
  cat0("Web Property Id:       ", x$webPropertyId)
  cat0("Upload ID:             ", x$id)
  cat0("Status:                ", x$status)
  
  
}}
\usage{
\method{print}{ga_custom_data_source_upload}(x, ...)
}
\description{
#' @export
print.ga_profileFilterLink <- function(x, ...){
  cat("==Google Analytics Profile (View) Filter Link==\n")
  cat0("profileFilterLink ID:  ", x$id)
  cat0("Rank:                  ", x$rank)
  cat0("Web Property Id:       ", x$webPropertyId)
  cat0("Upload ID:             ", x$id)
  cat0("Status:                ", x$status)
  
  
}
}
