% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print-method.R
\name{print.ga_model_result}
\alias{print.ga_model_result}
\title{#' @export
print.gar_FilterExpression <- function(x, ...){
  cat("==GA App+Web Filter Expression==\n")
  cat0("andGroup: ", lapply(x$andGroup, print)
  cat0("filter: ", x$filter)
  cat0("orGroup: ", lapply(x$orGroup, print)
  cat0("notExpression: ", x$notExpression)
}}
\usage{
\method{print}{ga_model_result}(x, ...)
}
\description{
#' @export
print.gar_Filter <- function(x, ...){
  cat("==GA App+Web Filter==\n")
  cat0("fieldName: ", x$fieldName)
  cat0("numericFilter: ", x$numericFilter)
  cat0("inListFilter: ", x$inListFilter)
  cat0("nullFilter: ", x$nullFilter)
  cat0("stringFilter: ", x$stringFilter)
  cat0("betweenFilter: ", x$betweenFilter)
}
}
\details{
#' @export
print.gar_StringFilter <- function(x, ...){
  cat("==GA App+Web StringFilter==\n")
  cat0("value: ", x$value)
  cat0("matchType: ", x$matchType)
  cat0("caseSensitive: ", x$caseSensitive)
}
}
