% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{rmNullObs}
\alias{rmNullObs}
\title{Recursively step down into list, removing all such objects}
\usage{
rmNullObs(x)
}
\description{
Recursively step down into list, removing all such objects
}
\keyword{internal}
