% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segments.R
\name{segment_element}
\alias{segment_element}
\title{Make a segment element}
\usage{
segment_element(name, operator = c("REGEXP", "BEGINS_WITH", "ENDS_WITH",
  "PARTIAL", "EXACT", "IN_LIST", "NUMERIC_LESS_THAN", "NUMERIC_GREATER_THAN",
  "NUMERIC_BETWEEN", "LESS_THAN", "GREATER_THAN", "EQUAL", "BETWEEN"),
  type = c("metric", "dimension"), not = FALSE, expressions = NULL,
  caseSensitive = NULL, minComparisonValue = NULL,
  maxComparisonValue = NULL, scope = c("SESSION", "USER", "HIT", "PRODUCT"),
  comparisonValue = NULL, matchType = c("PRECEDES", "IMMEDIATELY_PRECEDES"))
}
\arguments{
\item{name}{Name of the GA metric or dimension to segment on}

\item{operator}{How name shall operate on expression or comparisonValue}

\item{type}{A metric or dimension based segment element}

\item{not}{Should the element be the negation of what is defined}

\item{expressions}{[dim] What the name shall compare to}

\item{caseSensitive}{[dim] Whether to be case sensitive}

\item{minComparisonValue}{[dim] Minimum comparison values for BETWEEN}

\item{maxComparisonValue}{Max comparison value for BETWEEN operator}

\item{scope}{[met] Scope of the metric value}

\item{comparisonValue}{[met] What the name shall compare to}

\item{matchType}{If used in sequence segment, what behaviour}
}
\value{
An SegmentFilterClause object for use in a vector passed to \link{segment_vector}
}
\description{
Define the segment steps here, then combine into a vector and pass to \link{segment_vector}
  which in turn is combined into a list for \link{segment_ga4} 
  to finally be passed as a list to \link{google_analytics_4}
}
\seealso{
Other v4 segment functions: \code{\link{dynamicSegment}},
  \code{\link{orFiltersForSegment}},
  \code{\link{segmentDefinition}},
  \code{\link{segmentDimensionFilter}},
  \code{\link{segmentFilterClause}},
  \code{\link{segmentFilter}},
  \code{\link{segmentMetricFilter}},
  \code{\link{segmentObj_ga4}},
  \code{\link{segmentSequenceStep}},
  \code{\link{segment_define}}, \code{\link{segment_ga4}},
  \code{\link{segment_vector_sequence}},
  \code{\link{segment_vector_simple}},
  \code{\link{sequenceSegment}},
  \code{\link{simpleSegment}}
}

