% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ga_v4_get.R
\name{segment_filter}
\alias{segment_filter}
\title{Make a segment filter}
\usage{
segment_filter(type = c("metric", "dimensions"), matchType = c("PRECEDES",
  "IMMEDIATELY_PRECEDES"))
}
\arguments{
\item{matchType}{If to be used in sequence segment,
should it precede or immmediately precede}
}
\value{
An orFiltersForSegment object for use in segment_ga4
}
\description{
Make a segment filter
}
\seealso{
Other v4 segment functions: \code{\link{dynamicSegment}},
  \code{\link{orFiltersForSegment}},
  \code{\link{segmentDefinition}},
  \code{\link{segmentDimensionFilter}},
  \code{\link{segmentFilterClause}},
  \code{\link{segmentFilter}},
  \code{\link{segmentMetricFilter}},
  \code{\link{segmentObj_ga4}},
  \code{\link{segmentSequenceStep}},
  \code{\link{segment_ga4}}, \code{\link{sequenceSegment}},
  \code{\link{simpleSegment}}
}

