% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segments.R
\name{segment_vector_simple}
\alias{segment_vector_simple}
\title{Make a simple segment vector}
\usage{
segment_vector_simple(segment_elements)
}
\arguments{
\item{segment_elements}{A list of OR lists of \link{segment_element}}
}
\value{
A segment vector you can put in a list for use in \link{segment_ga4}
}
\description{
Make a simple segment vector
}
\seealso{
Other v4 segment functions: \code{\link{dynamicSegment}},
  \code{\link{orFiltersForSegment}},
  \code{\link{segmentDefinition}},
  \code{\link{segmentDimensionFilter}},
  \code{\link{segmentFilterClause}},
  \code{\link{segmentFilter}},
  \code{\link{segmentMetricFilter}},
  \code{\link{segmentObj_ga4}},
  \code{\link{segmentSequenceStep}},
  \code{\link{segment_define}},
  \code{\link{segment_element}}, \code{\link{segment_ga4}},
  \code{\link{segment_vector_sequence}},
  \code{\link{sequenceSegment}},
  \code{\link{simpleSegment}}
}

