% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny-modules-segments.R
\name{segmentBuilderUI}
\alias{segmentBuilderUI}
\title{Create a GAv4 Segment Builder}
\usage{
segmentBuilderUI(id)
}
\arguments{
\item{id}{Shiny id}
}
\value{
Shiny UI for use in app
}
\description{
Shiny Module for use with \link{segmentBuilder}
}
\examples{

\dontrun{

library(shiny)
library(googleAnalyticsR)

ui <- shinyUI(fluidPage(
               segmentBuilderUI("test1")
               ))
               
               
server <- shinyServer(function(input, output, session) {

segment <- callModule(segmentBuilder, "test1")

  .. use segment() in further gav4 calls.
  
})


# Run the application 
shinyApp(ui = ui, server = server)
}
}
