% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny-modules-segments.R
\name{segmentChain}
\alias{segmentChain}
\title{segmentChain}
\usage{
segmentChain(input, output, session, element_inputs)
}
\arguments{
\item{input}{shiny input}

\item{output}{shiny output}

\item{session}{shiny session}

\item{element_inputs}{from \link{segmentElementUI}}
}
\value{
Something
}
\description{
Shiny Module for use with \link{segmentChainUI}
}
\details{
Call via \code{shiny::callModule(segmentChain, "your_id")}
}
\keyword{internal}
