% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny-modules.R
\name{segmentChainSequence}
\alias{segmentChainSequence}
\title{segmentChainSequence}
\usage{
segmentChainSequence(input, output, session, segment_chain, element_inputs)
}
\arguments{
\item{input}{shiny input}

\item{output}{shiny output}

\item{session}{shiny session}

\item{segment_chain}{reactive segment_chain from \link{segmentChain}}
}
\value{
Something
}
\description{
Shiny Module for use with \link{segmentChainSequenceUI}
}
\details{
Call via \code{shiny::callModule(segmentChainSequence, "your_id")}
}

