% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny-modules.R
\name{segmentChainSequenceUI}
\alias{segmentChainSequenceUI}
\title{segmentChainSequence UI}
\usage{
segmentChainSequenceUI(id)
}
\arguments{
\item{id}{Shiny id}
}
\value{
Shiny UI
}
\description{
Shiny Module for use with \link{segmentChainSequence}
}

