% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny-modules-segments.R
\name{segmentChainUI}
\alias{segmentChainUI}
\title{segmentChain UI}
\usage{
segmentChainUI(id)
}
\arguments{
\item{id}{Shiny id}
}
\value{
Shiny UI
}
\description{
Shiny Module for use with \link{segmentChain}
}
\keyword{internal}
