% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny-modules.R
\name{segmentDefinitionBuilder}
\alias{segmentDefinitionBuilder}
\title{segmentDefinitionBuilder}
\usage{
segmentDefinitionBuilder(input, output, session, segment_sequence,
  element_inputs)
}
\arguments{
\item{input}{shiny input}

\item{output}{shiny output}

\item{session}{shiny session}

\item{segment_sequence}{segment sequence from \link{segmenChainSequence}}

\item{element_inputs}{element inputs}
}
\value{
Something
}
\description{
Shiny Module for use with \link{segmentDefinitionBuilderUI}
}
\details{
Call via \code{shiny::callModule(segmentDefinitionBuilder, "your_id")}
}

