% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny-modules.R
\name{segmentDefinitionBuilderUI}
\alias{segmentDefinitionBuilderUI}
\title{segmentDefinitionBuilder UI}
\usage{
segmentDefinitionBuilderUI(id)
}
\arguments{
\item{id}{Shiny id}
}
\value{
Shiny UI
}
\description{
Shiny Module for use with \link{segmentDefinitionBuilder}
}

