% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segments.R
\name{segmentDimensionFilter}
\alias{segmentDimensionFilter}
\title{segmentDimensionFilter}
\usage{
segmentDimensionFilter(name, expressions, operator = c("REGEXP",
  "BEGINS_WITH", "ENDS_WITH", "PARTIAL", "EXACT", "IN_LIST",
  "NUMERIC_LESS_THAN", "NUMERIC_GREATER_THAN", "NUMERIC_BETWEEN"),
  caseSensitive = FALSE, minComparisonValue = NULL,
  maxComparisonValue = NULL)
}
\description{
segmentDimensionFilter
}
\seealso{
Other v4 segment functions: \code{\link{dynamicSegment}},
  \code{\link{orFiltersForSegment}},
  \code{\link{segmentDefinition}},
  \code{\link{segmentFilterClause}},
  \code{\link{segmentFilter}},
  \code{\link{segmentMetricFilter}},
  \code{\link{segmentObj_ga4}},
  \code{\link{segmentSequenceStep}},
  \code{\link{segment_define}},
  \code{\link{segment_element}}, \code{\link{segment_ga4}},
  \code{\link{segment_vector_simple}},
  \code{\link{sequenceSegment}},
  \code{\link{simpleSegment}}
}

