% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny-modules-segments.R
\name{segmentElementUI}
\alias{segmentElementUI}
\title{A GAv4 segment element row}
\usage{
segmentElementUI(id)
}
\arguments{
\item{id}{shiny id}
}
\description{
A GAv4 segment element row
}
\keyword{internal}
