% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ga_v4_objects.R
\name{segmentFilter}
\alias{segmentFilter}
\title{Segment Filter}
\usage{
segmentFilter(matchComplement = FALSE, simpleSegment = NULL,
  sequenceSegment = NULL)
}
\arguments{
\item{sequenceSegment}{}
}
\description{
Segment Filter
}
\details{
If true, match the complement of simple or sequence segment above.
For example, to match all visits not from "New York", set to TRUE.
}
\seealso{
Other v4 segment functions: \code{\link{dynamicSegment}},
  \code{\link{orFiltersForSegment}},
  \code{\link{segmentDefinition}},
  \code{\link{segmentDimensionFilter}},
  \code{\link{segmentFilterClause}},
  \code{\link{segmentMetricFilter}},
  \code{\link{segmentObj_ga4}},
  \code{\link{segmentSequenceStep}},
  \code{\link{segment_filter}}, \code{\link{segment_ga4}},
  \code{\link{sequenceSegment}},
  \code{\link{simpleSegment}}
}

