% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ga_v4_objects.R
\name{segmentMetricFilter}
\alias{segmentMetricFilter}
\title{segmentMetricFilter}
\usage{
segmentMetricFilter(name, scope = c("PRODUCT", "HIT", "SESSION", "USER"),
  operator = c("LESS_THAN", "GREATER_THAN", "EQUAL", "BETWEEN"),
  comparisonValue = NULL, maxComparisonValue = NULL)
}
\description{
segmentMetricFilter
}
\seealso{
Other v4 segment functions: \code{\link{dynamicSegment}},
  \code{\link{orFiltersForSegment}},
  \code{\link{segmentDefinition}},
  \code{\link{segmentDimensionFilter}},
  \code{\link{segmentFilterClause}},
  \code{\link{segmentFilter}},
  \code{\link{segmentObj_ga4}},
  \code{\link{segmentSequenceStep}},
  \code{\link{segment_filter}}, \code{\link{segment_ga4}},
  \code{\link{sequenceSegment}},
  \code{\link{simpleSegment}}
}

