% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ga_v4_objects.R
\name{segmentObj_ga4}
\alias{segmentObj_ga4}
\title{Segment objects}
\usage{
segmentObj_ga4(dynamicSegment = NULL, segmentExpression = NULL)
}
\arguments{
\item{dynamicSegment}{A v4 segment object}

\item{segmentExpression}{A v3 segment string}
}
\value{
a list of class \code{segment_ga4}
}
\description{
Segment objects
}
\seealso{
Other v4 segment functions: \code{\link{dynamicSegment}},
  \code{\link{orFiltersForSegment}},
  \code{\link{segmentDefinition}},
  \code{\link{segmentDimensionFilter}},
  \code{\link{segmentFilterClause}},
  \code{\link{segmentFilter}},
  \code{\link{segmentMetricFilter}},
  \code{\link{segmentSequenceStep}},
  \code{\link{segment_filter}}, \code{\link{segment_ga4}},
  \code{\link{sequenceSegment}},
  \code{\link{simpleSegment}}
}

