% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segments.R
\name{sequenceSegment}
\alias{sequenceSegment}
\title{sequenceSegment}
\usage{
sequenceSegment(segmentSequenceStepList, firstStepMatch = FALSE)
}
\arguments{
\item{segmentSequenceStepList}{A list of \link{segmentSequenceStep}}

\item{firstStepMatch}{Should firstStep match the first hit?}
}
\value{
A sequenceSegment object
}
\description{
sequenceSegment
}
\seealso{
Other v4 segment functions: \code{\link{dynamicSegment}},
  \code{\link{orFiltersForSegment}},
  \code{\link{segmentDefinition}},
  \code{\link{segmentDimensionFilter}},
  \code{\link{segmentFilterClause}},
  \code{\link{segmentFilter}},
  \code{\link{segmentMetricFilter}},
  \code{\link{segmentObj_ga4}},
  \code{\link{segmentSequenceStep}},
  \code{\link{segment_define}},
  \code{\link{segment_element}},
  \code{\link{segment_vector_sequence}},
  \code{\link{segment_vector_simple}},
  \code{\link{simpleSegment}}
}
\keyword{internal}

