% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ga4_data_objects_filters.R
\name{StringFilter}
\alias{StringFilter}
\title{StringFilter Object}
\usage{
StringFilter(
  value,
  matchType = c("EXACT", "BEGINS_WITH", "ENDS_WITH", "CONTAINS", "FULL_REGEXP",
    "PARTIAL_REGEXP"),
  caseSensitive = TRUE
)
}
\arguments{
\item{value}{The string value used for the matching}

\item{matchType}{The match type for this filter}

\item{caseSensitive}{If true, the string value is case sensitive}
}
\value{
StringFilter object
}
\description{
StringFilter Object
}
\details{
The filter for string
}
\concept{StringFilter functions}
