% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{unitToList}
\alias{unitToList}
\title{Allow unit lists}
\usage{
unitToList(perhaps_list)
}
\arguments{
\item{perhaps_list}{A list or an element that will be wrapped in list()}
}
\value{
A list
}
\description{
If you need a list but only get one element, make it a list.
}
\details{
Makes it easier for users to use some functions by not worrying about list wrapping.
}
\keyword{internal}
