% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny-modules.R
\name{authDropdown}
\alias{authDropdown}
\title{authDropdown [Shiny Module]}
\usage{
authDropdown(input, output, session, ga.table)
}
\arguments{
\item{input}{shiny input}

\item{output}{shiny output}

\item{session}{shiny session}

\item{ga.table}{A table of GA tables}
}
\value{
GA View Id selected
}
\description{
Shiny Module for use with \link{authDropdownUI}
}
\details{
Call via \code{shiny::callModule(authDropdown, "your_id")}
}
\seealso{
Other Shiny modules: \code{\link{authDropdownUI}},
  \code{\link{multi_selectUI}}, \code{\link{multi_select}}
}
