% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny-modules.R
\name{authDropdownUI}
\alias{authDropdownUI}
\title{authDropdown UI}
\usage{
authDropdownUI(id, width = NULL, inColumns = FALSE)
}
\arguments{
\item{id}{Shiny id.}

\item{inColumns}{whether to wrap selectInputs in width=4 columns.

Shiny Module for use with \link{authDropdown}.}
}
\value{
Shiny UI
}
\description{
Makes a dropdown row for use for authentication.
}

