% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny-modules.R
\name{authDropdownUI}
\alias{authDropdownUI}
\title{authDropdown UI [Shiny Module]}
\usage{
authDropdownUI(id, width = NULL, inColumns = FALSE)
}
\arguments{
\item{id}{Shiny id.}

\item{width}{The width of the input}

\item{inColumns}{whether to wrap selectInputs in width=4 columns.

Shiny Module for use with \link{authDropdown}.}
}
\value{
Shiny UI
}
\description{
Makes a dropdown row for use for authentication.
}
\seealso{
Other Shiny modules: \code{\link{authDropdown}},
  \code{\link{multi_selectUI}}, \code{\link{multi_select}}
}
