% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filters.R
\name{dim_filter}
\alias{dim_filter}
\title{Make a dimension filter object}
\usage{
dim_filter(dimension, operator = c("REGEXP", "BEGINS_WITH", "ENDS_WITH",
  "PARTIAL", "EXACT", "NUMERIC_EQUAL", "NUMERIC_GREATER_THAN",
  "NUMERIC_LESS_THAN", "IN_LIST"), expressions, caseSensitive = FALSE,
  not = FALSE)
}
\arguments{
\item{dimension}{dimension name to filter on.}

\item{operator}{How to match the dimension.}

\item{expressions}{What to match.  A character vector if operator is "IN_LIST"}

\item{caseSensitive}{Boolean.}

\item{not}{Logical NOT operator. Boolean.}
}
\value{
An object of class \code{dim_fil_ga4} for use in \code{\link{filter_clause_ga4}}
}
\description{
Make a dimension filter object
}
\examples{

\dontrun{
library(googleAnalyticsR)

## authenticate, 
## or use the RStudio Addin "Google API Auth" with analytics scopes set
ga_auth()

## get your accounts
account_list <- google_analytics_account_list()

## pick a profile with data to query

ga_id <- account_list[23,'viewId']

## create filters on metrics
mf <- met_filter("bounces", "GREATER_THAN", 0)
mf2 <- met_filter("sessions", "GREATER", 2)

## create filters on dimensions
df <- dim_filter("source","BEGINS_WITH","1",not = TRUE)
df2 <- dim_filter("source","BEGINS_WITH","a",not = TRUE)

## construct filter objects
fc2 <- filter_clause_ga4(list(df, df2), operator = "AND")
fc <- filter_clause_ga4(list(mf, mf2), operator = "AND")

## make v4 request
ga_data1 <- google_analytics_4(ga_id, 
                               date_range = c("2015-07-30","2015-10-01"),
                               dimensions=c('source','medium'), 
                               metrics = c('sessions','bounces'), 
                               met_filters = fc, 
                               dim_filters = fc2, 
                               filtersExpression = "ga:source!=(direct)")
                               
}

}
\seealso{
Other filter functions: \code{\link{filter_clause_ga4}},
  \code{\link{met_filter}}
}
