% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segments.R
\name{dynamicSegment}
\alias{dynamicSegment}
\title{Dynamic Segment v4 object}
\usage{
dynamicSegment(name, userSegment, sessionSegment)
}
\arguments{
\item{name}{Name of dynamic segment}

\item{userSegment}{userSegment to include in the segment}

\item{sessionSegment}{sessionSegment to include in the segment}
}
\description{
Dynamic Segment v4 object
}
\seealso{
Other v4 segment functions: \code{\link{orFiltersForSegment}},
  \code{\link{segmentDefinition}},
  \code{\link{segmentDimensionFilter}},
  \code{\link{segmentFilterClause}},
  \code{\link{segmentFilter}},
  \code{\link{segmentMetricFilter}},
  \code{\link{segmentObj_ga4}},
  \code{\link{segmentSequenceStep}},
  \code{\link{segment_define}},
  \code{\link{segment_element}}, \code{\link{segment_ga4}},
  \code{\link{segment_vector_simple}},
  \code{\link{sequenceSegment}},
  \code{\link{simpleSegment}}
}

