% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segments.R
\name{dynamicSegment}
\alias{dynamicSegment}
\title{Dynamic Segment v4 object}
\usage{
dynamicSegment(name, userSegment, sessionSegment)
}
\arguments{
\item{name}{Name of dynamic segment}

\item{userSegment}{userSegment to include in the segment}

\item{sessionSegment}{sessionSegment to include in the segment}
}
\description{
Dynamic Segment v4 object
}
\keyword{internal}
