% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_web_objects.R, R/ga4_data_objects.R
\name{Entity}
\alias{Entity}
\title{Entity Object}
\usage{
Entity(propertyId)

Entity(propertyId)
}
\arguments{
\item{propertyId}{A Google Analytics App + Web property id}
}
\value{
Entity object

Entity object
}
\description{
Entity Object

Entity Object
}
\details{
The unique identifier of the property whose events are tracked.

The unique identifier of the property whose events are tracked.
}
\concept{Entity functions}
