% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ga_v4_get.R
\name{fetch_google_analytics_4}
\alias{fetch_google_analytics_4}
\title{Fetch multiple GAv4 requests}
\usage{
fetch_google_analytics_4(request_list)
}
\arguments{
\item{requests}{A list of requests created by \link{make_ga_4_req}}
}
\value{
A dataframe if one request, or a list of data.frames if multiple.
}
\description{
Fetch the GAv4 requests as created by \link{make_ga_4_req}
}

