% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_web_objects_filters.R
\name{FilterExpressionList}
\alias{FilterExpressionList}
\title{FilterExpressionList Object}
\usage{
FilterExpressionList(expressions)
}
\arguments{
\item{expressions}{A list of filter expressions}
}
\value{
FilterExpressionList object
}
\description{
FilterExpressionList Object
}
\details{
A list of filter expressions.
}
\concept{FilterExpressionList functions}
