% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adwords.R
\name{ga_adwords}
\alias{ga_adwords}
\title{Get AdWords Link meta data}
\usage{
ga_adwords(accountId, webPropertyId, webPropertyAdWordsLinkId)
}
\arguments{
\item{accountId}{Account Id}

\item{webPropertyId}{Web Property Id}

\item{webPropertyAdWordsLinkId}{AdWords Link Id}
}
\value{
AdWords Meta data
}
\description{
Get AdWords Link meta data
}
\seealso{
Other Google Ad management functions: 
\code{\link{ga_adwords_add_linkid}()},
\code{\link{ga_adwords_delete_linkid}()},
\code{\link{ga_adwords_list}()}
}
\concept{Google Ad management functions}
