% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auth.R, R/ga_v4_get.R
\name{ga_auth}
\alias{ga_auth}
\title{ga_auth}
\usage{
ga_auth(token = NULL, new_user = FALSE, verbose = TRUE)

ga_auth(token = NULL, new_user = FALSE, verbose = TRUE)
}
\arguments{
\item{token}{If you have an existing token already}

\item{new_user}{Go through the auth flow again}

\item{verbose}{More feedback}

\item{token}{If you have an existing token already}

\item{new_user}{Go through the auth flow again}

\item{verbose}{More feedback}
}
\description{
ga_auth

ga_auth
}

