% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_web_functions.R
\name{ga_aw_filter_expr}
\alias{ga_aw_filter_expr}
\title{Create a filter expression for use with App+Web Reports}
\usage{
ga_aw_filter_expr(..., type = c("default", "not", "and", "or"))
}
\arguments{
\item{type}{The type of filter}

\item{a_filter}{A Filter object or a FilterExpression or a list of FilterExpressions}
}
\description{
For use with \link{google_analytics_aw}
}
\details{
The fields in the same FilterExpression need to be either all dimensions or all metrics.
}
