% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user-activity.R
\name{ga_clientid_activity}
\alias{ga_clientid_activity}
\title{User Activity Request}
\usage{
ga_clientid_activity(ids, viewId, id_type = c("CLIENT_ID", "USER_ID"),
  activity_type = NULL, date_range = NULL)
}
\arguments{
\item{viewId}{The viewId}

\item{id_type}{Whether its userId or clientId}

\item{date_range}{A vector of start and end dates.  If not used will default to a week.}

\item{id}{The user or clientId.  You can send in a vector of them}
}
\description{
User Activity Request
}
\examples{

\dontrun{

googleAuthR::gar_set_client(scopes = "https://www.googleapis.com/auth/analytics")
ga_auth("test.oauth")
ga_clientid_activity(c("1106980347.1461227730", "476443645.1541099566"),
                     viewId = 81416156, 
                     date_range = c("2019-01-01","2019-02-01"))

}
}
\seealso{
https://developers.google.com/analytics/trusted-testing/user-reporting/
}
