% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user-activity.R
\name{ga_clientid_activity}
\alias{ga_clientid_activity}
\title{User Activity Request}
\usage{
ga_clientid_activity(viewId, id, id_type = c("CLIENT_ID", "USER_ID"),
  date_range = NULL)
}
\arguments{
\item{viewId}{The viewId}

\item{id}{The user or clientId}

\item{id_type}{Whether its userId or clientId}

\item{date_range}{A vector of start and end dates.  If not used will default to a week.}
}
\description{
User Activity Request
}
