% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user-activity.R
\name{ga_clientid_activity}
\alias{ga_clientid_activity}
\title{User Activity Request}
\usage{
ga_clientid_activity(ids, viewId, id_type = c("CLIENT_ID", "USER_ID"),
  activity_type = NULL, date_range = NULL)
}
\arguments{
\item{ids}{The userId or clientId.  You can send in a vector of them}

\item{viewId}{The viewId}

\item{id_type}{Whether its userId or clientId}

\item{activity_type}{If specified, filters down response to the activity type.  Choice between \code{"PAGEVIEW","SCREENVIEW","GOAL","ECOMMERCE","EVENT"}}

\item{date_range}{A vector of start and end dates.  If not used will default to a week.}
}
\value{
A list of data.frames: \code{$sessions} contains session level data. \code{$hits} contains individual activity data
}
\description{
Get activity on an individual user
}
\details{
The
}
\examples{

\dontrun{

# access data for individual users
uar <- ga_clientid_activity(c("1106980347.1461227730", "476443645.1541099566"),
                     viewId = 81416156, 
                     date_range = c("2019-01-01","2019-02-01"))

# access the data.frames returned:

# the session level data for the users passed in
uar$sessions

# the hit level activity for the users passed in
uar$hits
                     


}
}
\seealso{
\url{https://developers.google.com/analytics/devguides/reporting/core/v4/rest/v4/userActivity/search}
}
