% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user-activity.R
\name{ga_clientid_activity_unnest}
\alias{ga_clientid_activity_unnest}
\title{Unnest user activity columns}
\usage{
ga_clientid_activity_unnest(hits, column = c("customDimension", "ecommerce",
  "goals"))
}
\arguments{
\item{hits}{The hits data.frame with the columns to expand}

\item{column}{Which column to expand - one of \code{"customDimension","ecommerce","goals"}}
}
\description{
A function to help expand data out of nested columns
}
