% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom_data_source.R
\name{ga_custom_datasource}
\alias{ga_custom_datasource}
\title{List Custom Data Sources}
\usage{
ga_custom_datasource(accountId, webPropertyId)
}
\arguments{
\item{accountId}{Account Id}

\item{webPropertyId}{Web Property Id}
}
\value{
Custom Data Source
}
\description{
List Custom Data Sources
}

