% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom_data_source.R
\name{ga_custom_upload_list}
\alias{ga_custom_upload_list}
\title{List Custom Data Source Uploads}
\usage{
ga_custom_upload_list(accountId, webPropertyId, customDataSourceId)
}
\arguments{
\item{accountId}{Account Id}

\item{webPropertyId}{Web Property Id}

\item{customDataSourceId}{Custom data source Id}
}
\value{
Custom Data Source Uploads List
}
\description{
List Custom Data Source Uploads
}
\seealso{
Other custom datasource functions: \code{\link{ga_custom_datasource}},
  \code{\link{ga_custom_upload_file}},
  \code{\link{ga_custom_upload}}
}
