% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ga_data_order.R
\name{ga_data_order}
\alias{ga_data_order}
\title{Order DSL for GA4 OrderBy}
\usage{
ga_data_order(
  x,
  type = c("ALPHANUMERIC", "CASE_INSENSITIVE_ALPHANUMERIC", "NUMERIC")
)
}
\arguments{
\item{x}{Order DSL enabled syntax or the output of a previous call to this function - see examples}

\item{type}{Order Type - see details}
}
\value{
A \link{FilterExpression} object suitable for use in \link{ga_data}
}
\description{
Use with \link{ga_data} to create filters
}
\details{
The DSL rules are:


\itemize{
  \item{}{ Single filters can be used without wrapping in filter expressions}
}
}
\examples{


# session in descending order
ga_data_order(-sessions)

# city dimension in ascending alphanumeric order
ga_data_order(+city)

# as above plus sessions in descending order
ga_data_order(+city -sessions)

# as above plus activeUsers in ascending order
ga_data_order(+city -sessions +activeUsers)

# dayOfWeek dimension in ascending numeric order
ga_data_order(+dayOfWeek, type = "NUMERIC")


}
\seealso{
Other GA4 functions: 
\code{\link{ga_data_filter}()},
\code{\link{ga_data}()}
}
\concept{GA4 functions}
