% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filters.R
\name{ga_filter_apply_to_view}
\alias{ga_filter_apply_to_view}
\title{Apply an existing filter to view.}
\usage{
ga_filter_apply_to_view(filterId, accountId, propertyId, viewId)
}
\arguments{
\item{filterId}{The id of the filter to be addedd to profile/view}

\item{accountId}{Account Id of the account that contains the filter}

\item{propertyId}{Web property Id to create profile filter link for}

\item{viewId}{Profile/view Id to create profile filter link for}
}
\description{
Apply an existing filter to view.
}
\seealso{
Other managementAPI functions: \code{\link{ga_accounts}},
  \code{\link{ga_adwords_list}}, \code{\link{ga_adwords}},
  \code{\link{ga_custom_vars_list}},
  \code{\link{ga_custom_vars}},
  \code{\link{ga_experiment_list}},
  \code{\link{ga_experiment}}, \code{\link{ga_filter_add}},
  \code{\link{ga_filter_delete}},
  \code{\link{ga_filter_list}},
  \code{\link{ga_filter_update_filter_link}},
  \code{\link{ga_filter_update}},
  \code{\link{ga_filter_view_list}},
  \code{\link{ga_filter_view}}, \code{\link{ga_filter}},
  \code{\link{ga_goal_list}}, \code{\link{ga_goal}},
  \code{\link{ga_remarketing_estimate}},
  \code{\link{ga_remarketing_get}},
  \code{\link{ga_remarketing_list}},
  \code{\link{ga_segment_list}},
  \code{\link{ga_unsampled_list}},
  \code{\link{ga_unsampled}}, \code{\link{ga_users_list}},
  \code{\link{ga_view_list}}, \code{\link{ga_view}},
  \code{\link{ga_webproperty_list}},
  \code{\link{ga_webproperty}},
  \code{\link{google_analytics_account_list}}
}
