% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsl.R
\name{ga_filter_dsl}
\alias{ga_filter_dsl}
\title{Create filters via the DSL}
\usage{
ga_filter_dsl(x)
}
\arguments{
\item{x}{The DSL expression to evaluate into a GA filter}
}
\value{
A filter clause object
}
\description{
This lets you create filter objects using a convinient syntax, suitable to be passed to \link{google_analytics}
}
