% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_management.R
\name{ga_filter_view}
\alias{ga_filter_view}
\title{Get specific filter for view (profile)}
\usage{
ga_filter_view(accountId, webPropertyId, viewId, linkId)
}
\arguments{
\item{accountId}{Account Id}

\item{webPropertyId}{Web Property Id}

\item{viewId}{Profile Id}

\item{linkId}{Link Id}
}
\value{
filter list
}
\description{
Get specific filter for view (profile)
}
\seealso{
Other managementAPI functions: \code{\link{ga_adwords_list}},
  \code{\link{ga_adwords}}, \code{\link{ga_clientid_hash}},
  \code{\link{ga_experiment_list}},
  \code{\link{ga_experiment}}, \code{\link{ga_filter_add}},
  \code{\link{ga_filter_apply_to_view}},
  \code{\link{ga_filter_delete}},
  \code{\link{ga_filter_list}},
  \code{\link{ga_filter_update_filter_link}},
  \code{\link{ga_filter_update}},
  \code{\link{ga_filter_view_list}},
  \code{\link{ga_filter}}, \code{\link{ga_goal_add}},
  \code{\link{ga_goal_list}}, \code{\link{ga_goal_update}},
  \code{\link{ga_goal}},
  \code{\link{ga_remarketing_estimate}},
  \code{\link{ga_remarketing_get}},
  \code{\link{ga_remarketing_list}},
  \code{\link{ga_segment_list}},
  \code{\link{ga_unsampled_list}},
  \code{\link{ga_unsampled}}
}
\concept{managementAPI functions}
