% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta.R
\name{ga_meta}
\alias{ga_meta}
\title{Get current dimensions and metrics available in GA API.}
\usage{
ga_meta()
}
\value{
dataframe of dimensions and metrics available to use
}
\description{
Get current dimensions and metrics available in GA API.
}
\seealso{
\url{https://developers.google.com/analytics/devguides/reporting/metadata/v3/reference/metadata/columns/list}
}
