% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{ga_model}
\alias{ga_model}
\title{Use a model function created by \link{ga_model_make}}
\usage{
ga_model(viewId, model, load_libs = TRUE, ...)
}
\arguments{
\item{viewId}{The GA viewId to operate on}

\item{model}{A file location of a model object or a model object
created by \link{ga_model_make}}

\item{load_libs}{Whether to load the library requirements into your namespace}

\item{...}{Other arguments to pass into the model as needed}
}
\description{
Use a model function created by \link{ga_model_make}
}
\seealso{
Other GA modelling functions: 
\code{\link{ga_model_edit}()},
\code{\link{ga_model_example}()},
\code{\link{ga_model_load}()},
\code{\link{ga_model_make}()},
\code{\link{ga_model_save}()},
\code{\link{ga_model_write}()}
}
\concept{GA modelling functions}
