% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/platform.R
\name{ga_model}
\alias{ga_model}
\title{Use a model function created by ga_model_make}
\usage{
ga_model(viewId, model, ...)
}
\arguments{
\item{viewId}{The GA viewId to operate on}

\item{model}{A file location of a model object or a model object
created by \link{ga_model_make}}

\item{...}{Other arguments to pass to the model as needed}
}
\description{
Use a model function created by ga_model_make
}
\seealso{
Other GA modelling functions: \code{\link{ga_model_load}},
  \code{\link{ga_model_make}}, \code{\link{ga_model_save}}
}
\concept{GA modelling functions}
