% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/platform.R
\name{ga_model_edit}
\alias{ga_model_edit}
\title{Edit a created ga_model}
\usage{
ga_model_edit(model, data_f = NULL, required_columns = NULL,
  model_f = NULL, required_packages = NULL, description = NULL,
  outputShiny = NULL, renderShiny = NULL, renderShinyInput = NULL)
}
\arguments{
\item{model}{The model to edit - if a filepath will load model and 
save back edited model to the same file}

\item{data_f}{A function that gets the data -
must have viewId as first argument}

\item{required_columns}{What dimensions and metrics are required}

\item{model_f}{A function that inputs gadata, and outputs a list of assets -
must take data from result of data_f in first argument.}

\item{required_packages}{The packages needed for data_f and model_f to work}

\item{description}{An optional description of what the model does}

\item{outputShiny}{A shiny UI output function that will display the results renderShiny}

\item{renderShiny}{A shiny render function that will create the output for outputShiny from model_f}
}
\description{
Change features of a model by changing the functions within it.
}
\seealso{
Other GA modelling functions: \code{\link{ga_model_load}},
  \code{\link{ga_model_make}}, \code{\link{ga_model_save}},
  \code{\link{ga_model}}
}
\concept{GA modelling functions}
