% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/platform.R
\name{ga_model_example}
\alias{ga_model_example}
\title{Load an example model}
\usage{
ga_model_example(name, location = "googleAnalyticsR")
}
\arguments{
\item{name}{name of the model}

\item{location}{location of model}
}
\description{
Load an example model
}
\seealso{
Other GA modelling functions: \code{\link{ga_model_edit}},
  \code{\link{ga_model_load}}, \code{\link{ga_model_make}},
  \code{\link{ga_model_save}},
  \code{\link{ga_model_tweet}},
  \code{\link{ga_model_write}}, \code{\link{ga_model}}
}
\concept{GA modelling functions}
