% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/platform.R
\name{ga_model_load}
\alias{ga_model_load}
\title{Load a created model}
\usage{
ga_model_load(filename = "gamodel.rds")
}
\arguments{
\item{filename}{name to load model from}
}
\description{
Load a created model
}
\seealso{
Other GA modelling functions: \code{\link{ga_model_make}},
  \code{\link{ga_model_save}}, \code{\link{ga_model}}
}
\concept{GA modelling functions}
