% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/platform.R
\name{ga_model_save}
\alias{ga_model_save}
\title{Save a created model}
\usage{
ga_model_save(model, filename = "my-model.gamr")
}
\arguments{
\item{model}{model to save}

\item{filename}{name to save model under}
}
\description{
Save a created model
}
\seealso{
Other GA modelling functions: \code{\link{ga_model_edit}},
  \code{\link{ga_model_example}},
  \code{\link{ga_model_load}}, \code{\link{ga_model_make}},
  \code{\link{ga_model}}
}
