% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{ga_model_shiny}
\alias{ga_model_shiny}
\title{Create a Shiny app from a ga_model file}
\usage{
ga_model_shiny(
  model_location,
  template = ga_model_shiny_template("template1"),
  title = paste(basename(model_location), "ga_model", "googleAnalyticsR"),
  web_json = Sys.getenv("GAR_CLIENT_WEB_JSON"),
  scopes = "https://www.googleapis.com/auth/analytics.readonly",
  local_file = "",
  ...
)
}
\arguments{
\item{model_location}{The \link{ga_model} file location ("my_model.gamr")}

\item{template}{The template file for the Shiny app}

\item{title}{The title of the Shiny app}

\item{web_json}{The client.id json file for Web}

\item{scopes}{The scope the API requests will be under}

\item{local_file}{If not empty, will not launch Shiny app but write code to the file location you put here}

\item{...}{Extra variables the template may support}
}
\description{
Create a Shiny app from a ga_model file
}
