% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{ga_model_shiny_load}
\alias{ga_model_shiny_load}
\title{Load one model into a Shiny template}
\usage{
ga_model_shiny_load(model_n, ...)
}
\arguments{
\item{model_n}{The templated name of a model e.g. 'model1'}

\item{...}{Other arguments passed from shiny server}
}
\description{
Load one model into a Shiny template
}
