% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{ga_model_shiny_template}
\alias{ga_model_shiny_template}
\title{Get a Shiny template file}
\usage{
ga_model_shiny_template(name = "list")
}
\arguments{
\item{name}{the template name}
}
\description{
Gets a pre-created template from the googleAnalyticsR samples
}
