% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/platform.R
\name{ga_model_write}
\alias{ga_model_write}
\title{Write the ga_model functions to a file}
\usage{
ga_model_write(model, filepath = "ga_model.R")
}
\arguments{
\item{model}{The \code{ga_model} object to extract functions from to write}

\item{filepath}{The filepath to write the functions to}
}
\description{
Write the ga_model functions to a file
}
\seealso{
Other GA modelling functions: \code{\link{ga_model_edit}},
  \code{\link{ga_model_example}},
  \code{\link{ga_model_load}}, \code{\link{ga_model_make}},
  \code{\link{ga_model_save}},
  \code{\link{ga_model_tweet}}, \code{\link{ga_model}}
}
\concept{GA modelling functions}
