% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measurement_protocol.R
\name{ga_mp_send}
\alias{ga_mp_send}
\title{Make a Measurement Protocol v2 request}
\usage{
ga_mp_send(
  events,
  client_id,
  measurement_id,
  user_id = NULL,
  api_secret = Sys.getenv("GA_MP_SECRET"),
  debug = FALSE,
  timestamp_micros = NULL,
  user_properties = NULL,
  non_personalized_ads = TRUE
)
}
\arguments{
\item{events}{The events to send}

\item{client_id}{The client_id to associate with the event}

\item{measurement_id}{The measurement ID associated with a stream}

\item{user_id}{Optional. Unique id for the user}

\item{api_secret}{The secret generated in the GA4 UI - by default will look for environment arg \code{GA_MP_SECRET}}

\item{debug}{Send hits to the debug endpoint to validate hits}

\item{timestamp_micros}{Optional. A Unix timestamp (in microseconds) for the time to associate with the event.}

\item{user_properties}{Optional. The user properties for the measurement sent in as a named list.}

\item{non_personalized_ads}{Optional. Set to true to indicate these events should not be used for personalized ads.}
}
\value{
\code{TRUE} if successful, if \code{debug=TRUE} then validation messages if not a valid hit.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
Create a server side call to Google Analytics 4 via its Measurement Protocol
}
\details{
Create an API secret via \verb{Admin > Data Streams > choose your stream > Measurement Protocol > Create}

\code{user_id} can be used for \href{https://support.google.com/analytics/answer/9213390}{cross-platform analysis}

\code{timestamp_micros} should only be set to record events that happened in the past. This value can be overridden via user_property or event timestamps. Events can be backdated up to 48 hours. Note microseconds, not milliseconds.

\code{user_properties} - describe segments of your user base, such as language preference or geographic location.  See \href{https://developers.google.com/analytics/devguides/collection/protocol/ga4/user-properties?client_type=gtag}{User properties}

Ensure you also have user permission as specified in the \href{https://developers.google.com/analytics/devguides/collection/protocol/ga4/policy}{feature policy}

Invalid events are silently rejected with a 204 response, so use \code{debug=TRUE} to validate your events first.
}
\examples{
# preferably set this in .Renviron
Sys.setenv(GA_MP_SECRET="MY_SECRET")

# your GA4 settings
my_measurement_id <- "G-43MDXK6CLZ"
a_client_id <- 1234567

event <- ga_mp_event("an_event")
ga_mp_send(event, a_client_id, my_measurement_id, debug = TRUE)

another <- ga_mp_event("another_event")
ga_mp_send(list(event, another), 
           a_client_id, my_measurement_id, 
           debug = TRUE)
\dontrun{
# you can see sent events in the real-time reports
my_property_id <- 206670707
ga_data(my_property_id, 
        dimensions = "eventName", 
        metrics = "eventCount", 
        dimensionFilter = ga_data_filter(
           eventName == c("an_event","another_event")),
        realtime = TRUE)

}
}
\seealso{
\href{https://developers.google.com/analytics/devguides/collection/protocol/ga4}{Measurement Protocol (Google Analytics 4)}

Other Measurement Protocol functions: 
\code{\link{ga_mp_event_item}()},
\code{\link{ga_mp_event}()}
}
\concept{Measurement Protocol functions}
