% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remarketingAudiences.R
\name{ga_remarketing_estimate}
\alias{ga_remarketing_estimate}
\title{Estimate number of users added to the segment yesterday}
\usage{
ga_remarketing_estimate(remarketingAudience)
}
\arguments{
\item{remarketingAudience}{A remarketing audience object from \link{ga_remarketing_get}

Takes the segment definition from a remarketing audiences and runs 
  it against the viewId to see current estimated users

The total audience size is this figure for every membershipDurationDay from yesterday}
}
\value{
data.frame
}
\description{
Estimate number of users added to the segment yesterday
}
\seealso{
\href{https://support.google.com/analytics/answer/2611268?hl=en}{About remarketing audiences}

Other managementAPI functions: \code{\link{ga_accounts}},
  \code{\link{ga_adwords_list}}, \code{\link{ga_adwords}},
  \code{\link{ga_custom_vars_list}},
  \code{\link{ga_custom_vars}},
  \code{\link{ga_experiment_list}},
  \code{\link{ga_experiment}},
  \code{\link{ga_filter_list}},
  \code{\link{ga_filter_view_list}},
  \code{\link{ga_filter_view}}, \code{\link{ga_filter}},
  \code{\link{ga_goal_list}}, \code{\link{ga_goal}},
  \code{\link{ga_remarketing_get}},
  \code{\link{ga_remarketing_list}},
  \code{\link{ga_segment_list}},
  \code{\link{ga_unsampled_list}},
  \code{\link{ga_unsampled}}, \code{\link{ga_users_list}},
  \code{\link{ga_view_list}}, \code{\link{ga_view}},
  \code{\link{ga_webproperty_list}},
  \code{\link{ga_webproperty}},
  \code{\link{google_analytics_account_list}}
}
