% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measurement_protocol.R
\name{ga_trackme_event}
\alias{ga_trackme_event}
\title{Send a tracking hit for googleAnalyticsR package statistics}
\usage{
ga_trackme_event(debug = FALSE)
}
\arguments{
\item{debug}{Send as a debug event to see what will be sent}
}
\description{
This is the function that fires if you consent to tracking via \link{ga_trackme}
}
\details{
Running this function will send a Measurement Protocol hit via \link{ga_mp_send} only if the \code{.optin-googleanalyticsr} file is present in the same directory
}
\examples{

# this only works with a valid opt-in file present in the working directory
ga_trackme_event()
}
\seealso{
Other package tracking functions: 
\code{\link{ga_trackme}()}
}
\concept{package tracking functions}
