% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unsampled.R
\name{ga_unsampled}
\alias{ga_unsampled}
\title{Get Unsampled Report Meta Data}
\usage{
ga_unsampled(accountId, webPropertyId, profileId, unsampledReportId)
}
\arguments{
\item{accountId}{Account Id}

\item{webPropertyId}{Web Property Id}

\item{profileId}{Profile Id}

\item{unsampledReportId}{Unsampled Report Id}
}
\value{
Unsampled Report Meta Data
}
\description{
Get Unsampled Report Meta Data
}
\seealso{
Other unsampled download functions: 
\code{\link{ga_unsampled_download}()},
\code{\link{ga_unsampled_list}()}
}
\concept{unsampled download functions}
