% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/users.R
\name{ga_users_add}
\alias{ga_users_add}
\title{Create or update user access to Google Analytics}
\usage{
ga_users_add(email, permissions, accountId, webPropertyId = NULL,
  viewId = NULL)
}
\arguments{
\item{email}{The email of the user to add.  Has to have a Google account.}

\item{permissions}{Which permissions to add as a vector - \code{"MANAGE_USERS"},\code{"EDIT"},\code{"COLLABORATE"},\code{"READ_AND_ANALYZE"}}

\item{accountId}{Account Id}

\item{webPropertyId}{Web Property Id - set to NULL to operate on account level only}

\item{viewId}{viewId - set to NULL to operate on webProperty level only}
}
\value{
\code{TRUE} if successful
}
\description{
Create or update user access to Google Analytics
}
\seealso{
\href{https://developers.google.com/analytics/devguides/config/mgmt/v3/user-management}{Google help article on user permissions}

Other User management functions: \code{\link{ga_users_delete_linkid}},
  \code{\link{ga_users_delete}},
  \code{\link{ga_users_list}}
}
