% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/users.R
\name{ga_users_list}
\alias{ga_users_list}
\title{List Users}
\usage{
ga_users_list(accountId, webPropertyId = NULL, viewId = NULL)
}
\arguments{
\item{accountId}{Account Id}

\item{webPropertyId}{Web Property Id - leave NULL to fetch account level only}

\item{viewId}{viewId - leave NULL to fetch webProperty level only}
}
\description{
Will list users on an account, webproperty or view level
}
\seealso{
Other managementAPI functions: \code{\link{ga_accounts}},
  \code{\link{ga_adwords_list}}, \code{\link{ga_adwords}},
  \code{\link{ga_custom_vars_list}},
  \code{\link{ga_custom_vars}},
  \code{\link{ga_experiment_list}},
  \code{\link{ga_experiment}},
  \code{\link{ga_filter_list}},
  \code{\link{ga_filter_view_list}},
  \code{\link{ga_filter_view}}, \code{\link{ga_filter}},
  \code{\link{ga_goal_list}}, \code{\link{ga_goal}},
  \code{\link{ga_remarketing_estimate}},
  \code{\link{ga_remarketing_get}},
  \code{\link{ga_remarketing_list}},
  \code{\link{ga_segment_list}},
  \code{\link{ga_unsampled_list}},
  \code{\link{ga_unsampled}}, \code{\link{ga_view_list}},
  \code{\link{ga_view}}, \code{\link{ga_webproperty_list}},
  \code{\link{ga_webproperty}},
  \code{\link{google_analytics_account_list}}
}
