% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate.R
\name{getColNameOfClass}
\alias{getColNameOfClass}
\title{Gets the names of a dataframe's columns of a certain class}
\usage{
getColNameOfClass(df, class_name)
}
\arguments{
\item{df}{dataframe}

\item{class_name}{the R class to return columns of}
}
\value{
names of columns of class
}
\description{
Gets the names of a dataframe's columns of a certain class
}
