% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getData.R
\name{google_analytics}
\alias{google_analytics}
\title{Get Google Analytics v3 data}
\usage{
google_analytics(id, start, end, metrics = c("sessions", "bounceRate"),
  dimensions = NULL, sort = NULL, filters = NULL, segment = NULL,
  samplingLevel = c("DEFAULT", "FASTER", "HIGHER_PRECISION", "WALK"),
  max_results = 100, type = c("ga", "mcf"))
}
\arguments{
\item{id}{A character vector of View Ids to fetch from.}

\item{start}{Start date in YYY-MM-DD format.}

\item{end}{End date in YYY-MM-DD format.}

\item{metrics}{A character vector of metrics. With or without ga: prefix.}

\item{dimensions}{A character vector of dimensions. With or without ga: prefix.}

\item{sort}{How to sort the results, in form 'ga:sessions,-ga:bounceRate'}

\item{filters}{Filters for the result, in form 'ga:sessions>0;ga:pagePath=~blah'}

\item{segment}{How to segment.}

\item{samplingLevel}{Choose "WALK" to mitigate against sampling.}

\item{max_results}{Default 100. If greater than 10,000 then will batch GA calls.}

\item{type}{ga = Google Analytics v3; mcf = Multi-Channel Funels.}
}
\value{
For one id a data.frame of data, with meta-data in attributes.  
  For multiple id's, a list of dataframes.
}
\description{
Get Google Analytics v3 data
}
\seealso{
https://developers.google.com/analytics/devguides/reporting/core/v3/
}

