% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/getData.R
\name{google_analytics_account_list}
\alias{google_analytics_account_list}
\title{Get account list data including web property and View Id}
\usage{
google_analytics_account_list()
}
\value{
a dataframe of all account, webproperty and view data
}
\description{
Get account list data including web property and View Id
}
\seealso{
https://developers.google.com/analytics/devguides/config/mgmt/v3/mgmtReference/management/accountSummaries/list
}

