% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_web_functions.R
\name{google_analytics_aw}
\alias{google_analytics_aw}
\title{Google Analytics for GA4 (App+Web)}
\usage{
google_analytics_aw(
  propertyId,
  metrics,
  date_range,
  dimensions = NULL,
  dimensionFilter = NULL,
  dimensionDelimiter = "/",
  metricFilter = NULL,
  orderBys = NULL,
  limit = 100,
  realtime = FALSE
)
}
\arguments{
\item{propertyId}{A Google Analytics App + Web property id}

\item{metrics}{The metrics requested and displayed}

\item{date_range}{A vector of length two with start and end dates in YYYY-MM-DD format}

\item{dimensions}{The dimensions requested and displayed}

\item{dimensionFilter}{The filter clause of dimensions}

\item{dimensionDelimiter}{If combining dimensions in one column, the delimiter for the value field}

\item{metricFilter}{The filter clause of metrics}

\item{orderBys}{Specifies how rows are ordered in the response}

\item{limit}{The number of rows to return - use -1 to return all rows}

\item{realtime}{If TRUE then will call the real-time reports, that have a more limited set of dimensions/metrics - see \url{https://developers.google.com/analytics/devguides/reporting/data/v1/realtime-basics}}
}
\description{
Fetches Google Analytics data for Google Analytics 4 (Previously App+Web)
}
\details{

}
\examples{

\dontrun{

# send up to 4 date ranges
multi_date <- google_analytics_aw(
  206670707,
  metrics = c("activeUsers","sessions"),
  dimensions = c("date","city","dayOfWeek"),
  date_range = c("2020-03-31", "2020-04-27", "2020-04-30", "2020-05-27"),
  dimensionFilter = ga_aw_filter("city", "Copenhagen", "EXACT"),
  limit = 100
  )


# metric and dimension expressions

# create your own named metrics
met_expression <- google_analytics_aw(
  206670707,
  metrics = c("activeUsers","sessions",sessionsPerUser = "sessions/activeUsers"),
  dimensions = c("date","city","dayOfWeek"),
  date_range = c("2020-03-31", "2020-04-27"),
  limit = 100
  )

# create your own aggregation dimensions
dim_expression <- google_analytics_aw(
  206670707,
  metrics = c("activeUsers","sessions"),
  dimensions = c("date","city","dayOfWeek", cdow = "city/dayOfWeek"),
  date_range = c("2020-03-31", "2020-04-27"),
  limit = 100
  )
  
# run a real-time report (no date dimension allowed)
realtime <- google_analytics_aw(
  206670707,
  metrics = "activeUsers",
  dimensions = "city",
  dimensionFilter = ga_aw_filter("city", "Copenhagen", "EXACT"),
  limit = 100,
  realtime = TRUE)

}
}
\seealso{
\href{https://developers.google.com/analytics/devguides/reporting/data/v1}{Google Documentation}

Other BatchRunReportsRequest functions: 
\code{\link{BatchRunReportsRequest}()}
}
\concept{BatchRunReportsRequest functions}
