% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_web_functions.R
\name{google_analytics_aw}
\alias{google_analytics_aw}
\title{Google Analytics for App+Web}
\usage{
google_analytics_aw(
  propertyId,
  metrics,
  date_range,
  dimensions,
  dimensionFilter = NULL,
  metricFilter = NULL,
  orderBys = NULL,
  limit = 100,
  delimiter = "/"
)
}
\arguments{
\item{propertyId}{A Google Analytics App + Web property id}

\item{metrics}{The metrics requested and displayed}

\item{date_range}{A vector of length two with start and end dates in YYYY-MM-DD format}

\item{dimensions}{The dimensions requested and displayed}

\item{dimensionFilter}{The filter clause of dimensions}

\item{metricFilter}{The filter clause of metrics}

\item{orderBys}{Specifies how rows are ordered in the response}

\item{limit}{The number of rows to return}

\item{delimiter}{If combining dimensions in one column, the delimiter for the value field}
}
\description{
Fetches Google Analytics data for App+Web
}
\details{

}
\examples{

\dontrun{

# metric and dimension expressions
met_expression <- google_analytics_aw(
  206670707,
  metrics = c("activeUsers","sessions",sessionsPerUser = "sessions/activeUsers"),
  dimensions = c("date","city","dayOfWeek"),
  date_range = c("2020-03-31", "2020-04-27"),
  dimensionFilter = ga_aw_filter("city", "Copenhagen", "EXACT"),
  limit = 100
  )

dim_expression <- google_analytics_aw(
  206670707,
  metrics = c("activeUsers","sessions"),
  dimensions = c("date","city","dayOfWeek", cdow = "city/dayOfWeek"),
  date_range = c("2020-03-31", "2020-04-27"),
  dimensionFilter = ga_aw_filter("city", "Copenhagen", "EXACT"),
  limit = 100
  )

}
}
\seealso{
\href{https://developers.google.com/analytics/trusted-testing/analytics-data/}{Google Documentation}

Other BatchRunReportsRequest functions: 
\code{\link{BatchRunReportsRequest}()}
}
\concept{BatchRunReportsRequest functions}
