% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_web_functions.R
\name{google_analytics_aw}
\alias{google_analytics_aw}
\title{Returns multiple reports in a batch. All reports must be for the sameEntity.}
\usage{
google_analytics_aw(
  propertyId,
  metrics,
  date_range,
  dimensions,
  dimensionFilter = NULL,
  metricFilter = NULL,
  orderBys = NULL,
  limit = 100,
  delimiter = "/"
)
}
\arguments{
\item{BatchRunReportsRequest}{The \link{BatchRunReportsRequest} object to pass to this method
#' @importFrom googleAuthR gar_api_generator}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{

}
\examples{

\dontrun{

# metric and dimension expressions
met_expression <- google_analytics_aw(
  206670707,
  metrics = c("activeUsers","sessions",sessionsPerUser = "sessions/activeUsers"),
  dimensions = c("date","city","dayOfWeek"),
  date_range = c("2020-03-31", "2020-04-27"),
  dimensionFilter = ga_aw_filter("city", "Copenhagen", "EXACT"),
  limit = 100
  )

dim_expression <- google_analytics_aw(
  206670707,
  metrics = c("activeUsers","sessions"),
  dimensions = c("date","city","dayOfWeek", cdow = "city/dayOfWeek"),
  date_range = c("2020-03-31", "2020-04-27"),
  dimensionFilter = ga_aw_filter("city", "Copenhagen", "EXACT"),
  limit = 100
  )

}
}
\seealso{
\href{https://developers.google.com/analytics/trusted-testing/analytics-data/}{Google Documentation}

Other BatchRunReportsRequest functions: 
\code{\link{BatchRunReportsRequest}()}
}
\concept{BatchRunReportsRequest functions}
