% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_web_functions.R
\name{google_analytics_aw}
\alias{google_analytics_aw}
\title{Returns multiple reports in a batch. All reports must be for the sameEntity.}
\usage{
google_analytics_aw(
  propertyId,
  metrics,
  date_range,
  dimensions,
  dimensionFilter = NULL,
  metricFilter = NULL,
  orderBys = NULL,
  limit = 100
)
}
\arguments{
\item{BatchRunReportsRequest}{The \link{BatchRunReportsRequest} object to pass to this method
#' @importFrom googleAuthR gar_api_generator}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{

}
\seealso{
\href{https://developers.google.com/analytics/trusted-testing/analytics-data/}{Google Documentation}

Other BatchRunReportsRequest functions: 
\code{\link{BatchRunReportsRequest}()}
}
\concept{BatchRunReportsRequest functions}
