% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/googleAnalyticsR.R
\docType{package}
\name{googleAnalyticsR}
\alias{googleAnalyticsR}
\alias{googleAnalyticsR-package}
\title{Library for getting Google Analytics data into R}
\description{
docs: https://developers.google.com/analytics/trusted-testing/reporting/
}
\details{
https://developers.google.com/analytics/trusted-testing/reporting/rest/v4/reports/batchGet
}

