% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/googleAnalyticsR.R
\docType{package}
\name{googleAnalyticsR}
\alias{googleAnalyticsR}
\alias{googleAnalyticsR-package}
\title{Library for getting Google Analytics data into R}
\description{
https://developers.google.com/analytics/devguides/reporting/core/v4/rest/v4/reports/batchGet
}

