% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/googleAnalyticsR.R
\docType{package}
\name{googleAnalyticsR}
\alias{googleAnalyticsR}
\title{Library for getting Google Analytics data into R}
\description{
\if{html}{
  \figure{gar_logo_1000.png}{googleAnalyticsR}
}

Follow the online documentation here: \url{https://code.markedmondson.me/googleAnalyticsR/}
}
\details{
You may wish to set the below environment arguments for easier authentication

\verb{GA_CLIENT_ID GA_CLIENT_SECRET GA_WEB_CLIENT_ID GA_WEB_CLIENT_SECRET GA_AUTH_FILE}
}
