% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/googleAnalyticsR.R
\docType{package}
\name{googleAnalyticsR}
\alias{googleAnalyticsR}
\title{Library for getting Google Analytics data into R}
\description{
\ifelse{html}{
  \figure{logo.png}{googleAnalyticsR}
}{\strong{googleAnalyticsR}}

Follow the online documentation here: \url{https://code.markedmondson.me/googleAnalyticsR/}
}
\keyword{internal}
