% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cohorts.R
\name{make_cohort_group}
\alias{make_cohort_group}
\title{Create a cohort group}
\usage{
make_cohort_group(cohorts, lifetimeValue = FALSE, cohort_types = NULL)
}
\arguments{
\item{cohorts}{A named list of start/end date pairs}

\item{lifetimeValue}{lifetimeValue TRUE or FALSE.  Only works for webapps.}

\item{cohort_types}{placeholder, does nothing as only FIRST_VISIT_DATE supported.}
}
\value{
A cohortGroup object
}
\description{
Create a cohort group
}
\details{
Example: \code{list("cohort 1" = c("2015-08-01", "2015-08-01"), 
                    "cohort 2" = c("2015-07-01","2015-07-01"))}
}
\examples{

\dontrun{
library(googleAnalyticsR)

## authenticate, 
## or use the RStudio Addin "Google API Auth" with analytics scopes set
ga_auth()

## get your accounts
account_list <- google_analytics_account_list()

## pick a profile with data to query

ga_id <- account_list[23,'viewId']


## first make a cohort group

cohort4 <- make_cohort_group(list("cohort 1" = c("2015-08-01", "2015-08-01"), 
                                  "cohort 2" = c("2015-07-01","2015-07-01")))
                                  
## then call cohort report.  No date_range and must include metrics and dimensions
##   from the cohort list
cohort_example <- google_analytics_4(ga_id, 
                                     dimensions=c('cohort'), 
                                     cohort = cohort4, 
                                     metrics = c('cohortTotalUsers'))
                                     
}

}
\seealso{
\url{https://developers.google.com/analytics/devguides/reporting/core/v4/advanced#cohort_and_lifetime_value_ltv_dimensions_and_metrics}

Other v4 cohort functions: \code{\link{cohortGroup}},
  \code{\link{cohort_dimension_check}},
  \code{\link{cohort_metric_check}}, \code{\link{cohort}}
}
