% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_web_objects.R, R/ga4_data_objects.R
\name{Metric}
\alias{Metric}
\title{Metric Object}
\usage{
Metric(name, expression = NULL, invisible = NULL)

Metric(name, expression = NULL, invisible = NULL)
}
\arguments{
\item{name}{The name of the metric}

\item{expression}{A mathematical expression for derived metrics}

\item{invisible}{Indicates if a metric is invisible}
}
\value{
Metric object

Metric object
}
\description{
Metric Object

Metric Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The quantitative measurements of a report. For example, the metric eventCount is the total number of events. Requests are allowed up to 10 metrics.

Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The quantitative measurements of a report. For example, the metric eventCount is the total number of events. Requests are allowed up to 10 metrics.
}
\concept{Metric functions}
