% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ga_v4_objects.R
\name{order_type}
\alias{order_type}
\title{Make an OrderType object}
\usage{
order_type(field, sort_order = c("ASCENDING", "DESCENDING"),
  orderType = c("VALUE", "DELTA", "SMART", "HISTOGRAM_BUCKET",
  "DIMENSION_AS_INTEGER"))
}
\arguments{
\item{field}{One field to sort by}

\item{sort_order}{ASCENDING or DESCENDING}

\item{orderType}{Type of ordering}
}
\value{
A order_type_ga4 object for use in GAv4 fetch
}
\description{
Make an OrderType object
}
\details{
For multiple order sorting, create separate OrderType objects to pass
}
