% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_web_objects.R, R/ga4_data_objects.R
\name{OrderBy}
\alias{OrderBy}
\title{OrderBy Object}
\usage{
OrderBy(metric = NULL, pivot = NULL, dimension = NULL, desc = TRUE)

OrderBy(metric = NULL, pivot = NULL, dimension = NULL, desc = TRUE)
}
\arguments{
\item{metric}{Sorts results by a metric's values}

\item{pivot}{Sorts results by a metric's values within a pivot column group}

\item{dimension}{Sorts results by a dimension's values}

\item{desc}{If true, sorts by descending order}
}
\value{
OrderBy object

OrderBy object
}
\description{
OrderBy Object

OrderBy Object
}
\details{
The sort options.

The sort options.
}
\concept{OrderBy functions}
