% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pivots.R
\name{pivot_ga4_parse}
\alias{pivot_ga4_parse}
\title{pivot data parsing}
\usage{
pivot_ga4_parse(x, hasDateComparison)
}
\description{
pivot data parsing
}
\seealso{
Other pivot functions: \code{\link{makePivotData}},
  \code{\link{makePivotNames}}
}
\keyword{internal}
