% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ga_data_order.R, R/ga_data_pivot.R
\name{PivotOrderBy}
\alias{PivotOrderBy}
\title{PivotOrderBy Object}
\usage{
PivotOrderBy(pivotSelections = NULL, metricName = NULL)

PivotOrderBy(pivotSelections = NULL, metricName = NULL)
}
\arguments{
\item{pivotSelections}{Used to select a dimension name and value pivot}

\item{metricName}{In the response to order by, order rows by this column}
}
\value{
PivotOrderBy object

PivotOrderBy object
}
\description{
PivotOrderBy Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Sorts by a pivot column group.

Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Sorts by a pivot column group.
}
\concept{PivotOrderBy functions}
