% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segments.R
\name{segment_ga4}
\alias{segment_ga4}
\title{Make a segment object for use}
\usage{
segment_ga4(name, segment_id = NULL, user_segment = NULL,
  session_segment = NULL)
}
\arguments{
\item{name}{The name of the segment for the reports.}

\item{segment_id}{The segment ID of a built in or custom segment e.g. gaid::-3}

\item{user_segment}{A list of \code{segment_define}'s that apply to users}

\item{session_segment}{A list of \code{segment_define}'s that apply to sessions}
}
\value{
a segmentFilter object. You can pass a list of these to the request.
}
\description{
Make a segment object for use
}
\seealso{
Other v4 segment functions: \code{\link{dynamicSegment}},
  \code{\link{orFiltersForSegment}},
  \code{\link{segmentDefinition}},
  \code{\link{segmentDimensionFilter}},
  \code{\link{segmentFilterClause}},
  \code{\link{segmentFilter}},
  \code{\link{segmentMetricFilter}},
  \code{\link{segmentObj_ga4}},
  \code{\link{segmentSequenceStep}},
  \code{\link{segment_define}},
  \code{\link{segment_element}},
  \code{\link{segment_vector_simple}},
  \code{\link{sequenceSegment}},
  \code{\link{simpleSegment}}
}

