% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ga_v4_get.R
\name{segment_ga4}
\alias{segment_ga4}
\title{Make a segment object for use}
\usage{
segment_ga4(name, segment_filters, exclude = FALSE, type = c("simple",
  "sequence"), firstStepMatch = FALSE)
}
\arguments{
\item{name}{The name of the segment for the reports.}

\item{segment_filters}{A vector of \code{segment_filter}.}

\item{exclude}{Should this match or exclude the segment defined.}

\item{type}{Simple or Sequence segment.}

\item{firstStepMatch}{If a sequence segment, should first step match first hit.}
}
\value{
a segmentFilter object. You can pass a list of these to the request.
}
\description{
Make a segment object for use
}
\seealso{
Other v4 segment functions: \code{\link{dynamicSegment}},
  \code{\link{orFiltersForSegment}},
  \code{\link{segmentDefinition}},
  \code{\link{segmentDimensionFilter}},
  \code{\link{segmentFilterClause}},
  \code{\link{segmentFilter}},
  \code{\link{segmentMetricFilter}},
  \code{\link{segmentObj_ga4}},
  \code{\link{segmentSequenceStep}},
  \code{\link{segment_filter}},
  \code{\link{sequenceSegment}},
  \code{\link{simpleSegment}}
}

