% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segments.R
\name{segment_vector_sequence}
\alias{segment_vector_sequence}
\title{Make sequenceSegment}
\usage{
segment_vector_sequence(segment_elements, stepMatchList = NULL,
  firstStepMatch = FALSE)
}
\arguments{
\item{stepMatchList}{a list same length as segment_elements of
c("PRECEDES", "IMMEDIATELY_PRECEDES") for each step. 
If omitted, assumed all to be "PRECEDES"}

\item{firstStepMatch}{FALSE default}

\item{segment_element_list}{a list of lists of segment elements}
}
\description{
Make sequenceSegment
}

