% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segments.R
\name{segment_vector_sequence}
\alias{segment_vector_sequence}
\title{Make sequenceSegment}
\usage{
segment_vector_sequence(segment_elements, firstStepMatch = FALSE)
}
\arguments{
\item{segment_elements}{a list of OR lists of segment elements}

\item{firstStepMatch}{FALSE default}
}
\description{
\code{segment_vector_sequence} is in the hierarchy of segment creation, for which you will also need:
\itemize{
 \item \link{segment_define} : AND combination of segmentFilters
 \item \link{segment_vector_simple} or \link{segment_vector_sequence}
 \item \link{segment_element} that are combined in OR lists for \code{segment_vectors_*}
}
}
\seealso{
Other v4 segment functions: \code{\link{segment_define}},
  \code{\link{segment_element}}, \code{\link{segment_ga4}},
  \code{\link{segment_vector_simple}}
}
