% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segments.R
\name{segmentDefinition}
\alias{segmentDefinition}
\title{Segment Definition}
\usage{
segmentDefinition(segmentFilterList)
}
\arguments{
\item{segmentFilterList}{A list of segmentFilters

SegmentDefinition defines the segment to be a set of 
  SegmentFilters which are combined together with a logical AND operation.}
}
\description{
Segment Definition
}
\seealso{
Other v4 segment functions: \code{\link{dynamicSegment}},
  \code{\link{orFiltersForSegment}},
  \code{\link{segmentDimensionFilter}},
  \code{\link{segmentFilterClause}},
  \code{\link{segmentFilter}},
  \code{\link{segmentMetricFilter}},
  \code{\link{segmentObj_ga4}},
  \code{\link{segmentSequenceStep}},
  \code{\link{segment_define}},
  \code{\link{segment_element}}, \code{\link{segment_ga4}},
  \code{\link{segment_vector_sequence}},
  \code{\link{segment_vector_simple}},
  \code{\link{sequenceSegment}},
  \code{\link{simpleSegment}}
}
\keyword{internal}

