% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segments.R
\name{segmentDefinition}
\alias{segmentDefinition}
\title{Segment Definition}
\usage{
segmentDefinition(segmentFilterList)
}
\arguments{
\item{segmentFilterList}{A list of segmentFilters

SegmentDefinition defines the segment to be a set of 
  SegmentFilters which are combined together with a logical AND operation.}
}
\description{
Segment Definition
}
\keyword{internal}
