% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segments.R
\name{segmentFilter}
\alias{segmentFilter}
\title{Segment Filter}
\usage{
segmentFilter(not = FALSE, simpleSegment = NULL, sequenceSegment = NULL)
}
\arguments{
\item{not}{Negation of this filter}

\item{simpleSegment}{An simple segment}

\item{sequenceSegment}{An sequence segment}
}
\description{
Segment Filter
}
\details{
If true, match the complement of simple or sequence segment above.
For example, to match all visits not from "New York", set to TRUE.
}
\seealso{
Other v4 segment functions: \code{\link{dynamicSegment}},
  \code{\link{orFiltersForSegment}},
  \code{\link{segmentDefinition}},
  \code{\link{segmentDimensionFilter}},
  \code{\link{segmentFilterClause}},
  \code{\link{segmentMetricFilter}},
  \code{\link{segmentObj_ga4}},
  \code{\link{segmentSequenceStep}},
  \code{\link{segment_define}},
  \code{\link{segment_element}}, \code{\link{segment_ga4}},
  \code{\link{segment_vector_sequence}},
  \code{\link{segment_vector_simple}},
  \code{\link{sequenceSegment}},
  \code{\link{simpleSegment}}
}
\keyword{internal}

