% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segments.R
\name{segmentFilterClause}
\alias{segmentFilterClause}
\title{segmentFilterClause}
\usage{
segmentFilterClause(not = FALSE, dimensionFilter = NULL,
  metricFilter = NULL)
}
\description{
Make this internal
}
\keyword{internal}
