% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segments.R
\name{segmentMetricFilter}
\alias{segmentMetricFilter}
\title{segmentMetricFilter}
\usage{
segmentMetricFilter(name, scope = c("PRODUCT", "HIT", "SESSION", "USER"),
  operator = c("LESS_THAN", "GREATER_THAN", "EQUAL", "BETWEEN"),
  comparisonValue = NULL, maxComparisonValue = NULL)
}
\description{
segmentMetricFilter
}
\keyword{internal}
