% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segments.R
\name{segmentObj_ga4}
\alias{segmentObj_ga4}
\title{Segment objects}
\usage{
segmentObj_ga4(dynamicSegment = NULL, segmentId = NULL)
}
\arguments{
\item{dynamicSegment}{A v4 segment object}

\item{segmentId}{A v3 segment string}
}
\value{
a list of class \code{segment_ga4}
}
\description{
Segment objects
}
\keyword{internal}
