% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segments.R
\name{segmentSequenceStep}
\alias{segmentSequenceStep}
\title{segmentSequenceStep}
\usage{
segmentSequenceStep(orFiltersForSegmentList, matchType = c("PRECEDES",
  "IMMEDIATELY_PRECEDES"))
}
\arguments{
\item{orFiltersForSegmentList}{A list of \link{orFiltersForSegment}}

\item{matchType}{Should it precede or immediatly precede the next step}
}
\value{
segmentSequenceStep object
}
\description{
segmentSequenceStep
}
\seealso{
Other v4 segment functions: \code{\link{dynamicSegment}},
  \code{\link{orFiltersForSegment}},
  \code{\link{segmentDefinition}},
  \code{\link{segmentDimensionFilter}},
  \code{\link{segmentFilterClause}},
  \code{\link{segmentFilter}},
  \code{\link{segmentMetricFilter}},
  \code{\link{segmentObj_ga4}},
  \code{\link{segment_define}},
  \code{\link{segment_element}}, \code{\link{segment_ga4}},
  \code{\link{segment_vector_sequence}},
  \code{\link{segment_vector_simple}},
  \code{\link{sequenceSegment}},
  \code{\link{simpleSegment}}
}
\keyword{internal}

