% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segments.R
\name{segmentSequenceStep}
\alias{segmentSequenceStep}
\title{segmentSequenceStep}
\usage{
segmentSequenceStep(orFiltersForSegmentList, matchType = c("PRECEDES",
  "IMMEDIATELY_PRECEDES"))
}
\arguments{
\item{orFiltersForSegmentList}{A list of \link{orFiltersForSegment}}

\item{matchType}{Should it precede or immediately precede the next step}
}
\value{
segmentSequenceStep object
}
\description{
segmentSequenceStep
}
\keyword{internal}
