% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segments.R
\name{sequenceSegment}
\alias{sequenceSegment}
\title{sequenceSegment}
\usage{
sequenceSegment(segmentSequenceStepList, firstStepMatch = FALSE)
}
\arguments{
\item{segmentSequenceStepList}{A list of \link{segmentSequenceStep}}

\item{firstStepMatch}{Should firstStep match the first hit?}
}
\value{
A sequenceSegment object
}
\description{
sequenceSegment
}
\keyword{internal}
